refactor(core): split embedding modules
