+++ .lore.md
@@ -41,7 +41,7 @@
 * **embedding-units 64-chunk cap**: Chose structural \`"\n\x1f"\` boundaries and a 64-chunk temporal cap over splitting on tool-looking text, dropping overflow, or allowing unlimited vec0 rows because tool output may mimic headers and unbounded fan-out expands storage and KNN widening. Keep units dependency-free; preserve prose/reasoning verbatim and reduced tool envelopes. Units \`0..62\` stay separate, while index \`63+\` is joined into the final chunk so evidence is retained. Embed through token-area-bounded \`embedInTokenBatches(..., "document")\`, then atomically store the complete vector set.
 
 <!-- lore:01a05e22-e32b-701d-914e-9a031caa54bd -->
-* **EmbeddingProviderError privacy boundary**: Chose one fixed, cause-free \`EmbeddingProviderError\` at remote, worker, and custom-provider boundaries over rethrowing raw failures because fetch diagnostics, malformed bodies, getters, and custom errors can contain provider-controlled content. Keep its own properties non-enumerable; \`String(error)\`, serialization, and \`cause\` must expose no diagnostics. Voyage may expose HTTP status, but never response bodies or \`statusText\`.
+* **EmbeddingProviderError privacy boundary**: Chose one fixed, cause-free \`EmbeddingProviderError\` at remote, worker, and custom-provider boundaries over rethrowing raw failures because fetch diagnostics, malformed bodies, getters, and custom errors can contain provider-controlled content. At the shared \`embedWithProvider\` chokepoint, reconstruct allowlisted control errors instead of returning provider-owned objects, and catch hostile \`Proxy\`/\`instanceof\` inspection failures. Keep error properties non-enumerable; \`String(error)\`, serialization, and \`cause\` must expose no diagnostics. Voyage may expose HTTP status, but never response bodies or \`statusText\`.
 
 <!-- lore:01a05a11-c9de-7235-9324-5c8b17c720f7 -->
 * **ensureVec0Store per-table repair**: Chose savepoint-atomic per-table vec0 repair over rebuilding all four indexes because one missing or malformed derived table must not delete compatible sibling vectors. Preserve recognized same-width tables, recreate only missing/incompatible siblings, infer width from healthy physical DDL, persist metadata, and roll back the entire repair on failure; genuine configured-width changes still rebuild every fixed-width index.
@@ -137,7 +137,7 @@
 * **Sync payload and versioning invariants**: Push rows as toRemoteRow(pickSyncColumns(table,row)), with is\_deleted:false and content\_hash/revision only for versioned tables. Never push scope\_id; the server fills it from auth.uid(). Conflict targets are scope\_id plus table ID columns. Strip remote-only columns before local apply and local-only columns before push. Soft-delete with update({is\_deleted:true}).match(decomposeId(...)). SyncTableMeta.versioned defaults true; knowledge\_entity\_refs is non-versioned and must omit content\_hash/revision or PGRST204 leaves the push cursor stuck.
 
 <!-- lore:019f1fde-f17e-7835-a964-dfb1b4c697f5 -->
-* **Temporal backfill queue admission**: PR #1111 (\`fix/temporal-backfill-pool-gate\`) replaces the session-activity idle-gate (PR #1108) with an embed-pool capacity gate. Root cause of old gate: 5 active sessions kept backfill parked continuously; cursor never advanced during 8+ min window. New signal: module-level \`\_recallEmbedsInFlight\` counter in \`embedding.ts\`. Incremented in \`embed()\` when \`inputType === "query" && texts.length === 1\`; decremented in \`finally\`. Exported as \`recallEmbedsInFlight()\` + \`\_setRecallEmbedsInFlightForTest()\` seam. Gate factory \`makeTemporalBackfillGate(deps)\` parks when \`isPaused() || isEmbedBusy()\`. Wired in \`pipeline.ts\` as \`isEmbedBusy: () => embedding.recallEmbedsInFlight() > 0\`. Backfill always uses \`"document"\` inputType → \`isRecall\` always false → no self-gating. Throwing gate predicate caught with \`break\` in \`awaitBackfillIdle\` — never wedges walk.
+* **Temporal backfill queue admission**: PR #1111 (\`fix/temporal-backfill-pool-gate\`) replaces the session-activity idle-gate (PR #1108) with an embed-pool capacity gate. Root cause of old gate: 5 active sessions kept backfill parked continuously; cursor never advanced during 8+ min window. Core owns this invariant so every host gets it: \`embedding/runtime.ts\` increments \`\_recallEmbedsInFlight\` in \`embed()\` when \`inputType === "query" && texts.length === 1\` and decrements it in \`finally\`; \`embedding/backfill.ts\` parks temporal work while \`recallEmbedsInFlight() > 0\`, additionally composing the host's optional \`shouldPause\` policy. The gateway therefore supplies only \`isBackgroundPaused()\`. Backfill always uses \`"document"\` input type, so it never self-gates. A throwing host predicate is caught with \`break\` in \`awaitBackfillIdle\`, preventing a wedged walk.
 
 <!-- lore:01a042a2-c98e-76d5-9cc8-8064f2e62130 -->
 * **temporal\_messages sync subset**: Chose syncing only \`temporal\_messages.id ∈ ⋃ distillations.source\_ids\` over syncing every temporal message because undistilled observations must never leave the device. The table is append-only, unversioned, has no capture trigger, is enqueued by distillation fanout, and encrypts content/metadata on the wire. \`distilled\` and \`restored\_at\` remain local-only: remote inserts are archival, while conflicts preserve native residency state.
