+++ packages/core/src/embedding.ts
@@ -1,92 +1,105 @@
 /**
- * Embedding integration for vector search.
+ * Public embedding facade.
  *
- * Supports multiple embedding providers behind a common interface:
- *   - "local" (default): @huggingface/transformers + nomic-embed-text-v1.5
- *     (768 dims, Matryoshka-capable). Runs ONNX inference in a worker thread.
- *   - "voyage": Voyage AI API (voyage-code-3, 1024 dims)
- *   - "openai": OpenAI API (text-embedding-3-small, 1536 dims)
- *
- * Provides embedding generation, pure-JS cosine similarity, and vector search
- * over the knowledge and distillation tables. All operations are gated behind
- * `search.embeddings.enabled` config + the provider's API key env var — falls
- * back silently to FTS-only when unavailable.
+ * Implementation is split by responsibility so lifecycle ownership, search,
+ * document writes, and migrations/backfills can evolve independently without
+ * changing the long-standing public API imported by core and gateway callers.
  */
 
-import { freemem } from "node:os";
-import { createHash, randomUUID } from "node:crypto";
-import { performance } from "node:perf_hooks";
-import {
-  databaseInTransaction,
-  db,
-  getKV,
-  isCurrentDatabase,
-  setKV,
-  withSavepoint,
-  withTransaction,
-} from "./db";
-import { isVecAvailable } from "./db/vec";
-import {
-  clearAllEmbeddings,
-  copyBlobsToVec0,
-  dropEmbeddingColumn,
-  type EmbeddingTable,
-  embeddingColumnExists,
-  ensureVec0Store,
-  hasEmbeddingSql,
-  missingEmbeddingSql,
-  readStorageMode,
-  resolveReadMode,
-  setStorageMode,
-  storeEmbedding,
-  TEMPORAL_PARTITION_MODE_KEY,
-} from "./db/vec-store";
-import { config } from "./config";
-import * as log from "./log";
-import { recordVecReadLatency } from "./vec-latency";
-import { vendorModelInfo } from "./embedding-vendor";
-import { nativeIntraOpThreads } from "./ort-native";
-import {
-  MIN_EMBED_TOKENS,
-  MODEL_MAX_TOKENS,
-  PER_WORKER_MEM_BUDGET_BYTES,
-  EMBED_POOL_ABS_MAX,
-  backoffEmbedCap,
-  clampFreeToContainerLimit,
-  desiredEmbedPoolSize,
-  memoryModelEmbedCap,
-  reconcileEmbedCap,
-  reprobeEmbedCap,
-  shouldReprobeEmbedCap,
-  type PersistedEmbedCap,
-} from "./embedding-cap";
-import {
-  EMBED_OOM_EXIT_CODE,
-  isMissingLocalStackError,
-  isWasmFatalError,
-  type EmbedRequest,
-  type WorkerInbound,
-  type WorkerOutbound,
-  type WorkerInitData,
-} from "./embedding-worker-types";
-import {
-  runVectorQuery,
-  type DistillationVectorHit,
-  type VectorHit,
-  type VectorQuerySpec,
-} from "./vector-query";
-import { tryPoolVectorSearch, VECTOR_SEARCH_TIMED_OUT } from "./vector-pool";
-import { currentTenantId } from "./tenant";
-import {
-  enqueueTemporalEmbedding,
-  invalidateTemporalEmbedding,
-} from "./temporal-embedding-admission";
-import { TEMPORAL_EMBEDDING_MIN_CONTENT_LENGTH } from "./embedding-units";
+export {
+  EmbeddingAbortError,
+  EmbeddingQueueCapacityError,
+  EmbeddingRequestAbortedError,
+  EmbeddingWorkerWatchdogError,
+  LocalProviderUnavailableError,
+  awaitWorkerShutdown,
+  l2Normalize,
+} from "./embedding/contract";
+export type {
+  EmbeddingAbortCode,
+  EmbeddingAbortPhase,
+  EmbeddingOperationOptions,
+  EmbeddingProvider,
+  ShutdownableWorker,
+} from "./embedding/contract";
 
-// The cosine/BLOB helpers moved to ./vector-query (a leaf module the read
-// worker can import without pulling in the provider chain). Re-exported here so
-// existing `embedding.toBlob` / `embedding.cosineSimilarity` / `embedding.fromBlob`
-// call sites across the codebase keep working unchanged.
+export {
+  _getLocalInitRetryAtForTest,
+  _markLocalProviderUnavailable,
+  _persistEmbedCap,
+  _readPersistedEmbedCap,
+  _resetLocalProviderProbe,
+  _setConstrainedMemoryForTest,
+  _setContainerFreeForTest,
+  _setLocalInitCooldownMsForTest,
+  _setLocalInitRetryAtForTest,
+  _setTestWorkerFactory,
+  computeInitRetryDelayMs,
+  setEmbeddingFailureHook,
+} from "./embedding/local";
+export type { EmbeddingFailureInfo } from "./embedding/local";
+
+export {
+  _configuredBackfillCpuDuty,
+  _configuredEmbedPoolSize,
+  _setEmbedPoolSizeForTest,
+  _setEmbeddingWorkerWatchdogsForTest,
+  _setPoolFreememForTest,
+} from "./embedding/pool";
+
+export {
+  _resetProviderShutdownTrackingForTest,
+  _restoreProvider,
+  _saveAndClearProvider,
+  _setRecallEmbedsInFlightForTest,
+  _shutdownAndDisable,
+  embed,
+  embeddingStatus,
+  ensureEmbeddingReady,
+  isAvailable,
+  maybeSelfHealEmbeddingProvider,
+  recallEmbedsInFlight,
+  resetProvider,
+  shutdownProvider,
+} from "./embedding/runtime";
+export type {
+  EmbeddingHealth,
+  EmbeddingHealthState,
+} from "./embedding/runtime";
+
+export { pickRemoteFallback } from "./embedding/remote";
+
+export {
+  embedDistillation,
+  embedEntity,
+  embedInTokenBatches,
+  embedKnowledgeEntry,
+  settleDocumentEmbeds,
+  warmupEmbedding,
+} from "./embedding/documents";
+export { MAX_TEMPORAL_CHUNKS_PER_MESSAGE } from "./embedding-units";
+
+export {
+  vectorSearch,
+  vectorSearchAllDistillations,
+  vectorSearchDistillations,
+  vectorSearchEntities,
+  vectorSearchTemporal,
+} from "./embedding/search";
+
+export {
+  backfillDistillationEmbeddings,
+  backfillEmbeddings,
+  backfillEntityEmbeddings,
+  backfillTemporalEmbeddings,
+  checkConfigChange,
+  formatTemporalRechunkProgress,
+  maybeCutoverToVec0,
+  resetTemporalRechunkProgress,
+  runStartupBackfill,
+} from "./embedding/backfill";
+export type { BackfillOptions, BackfillStats } from "./embedding/backfill";
+
 export {
   cosineSimilarity,
   fromBlob,
@@ -94,4788 +107,3 @@ export {
   type DistillationVectorHit,
   type VectorHit,
 } from "./vector-query";
-
-/** Timeout for embedding API fetch calls (ms). Prevents a hanging API from
- *  blocking the recall tool indefinitely. 10s is generous for typical 100-500ms
- *  embedding calls but bounded enough to avoid minutes-long hangs. */
-const EMBED_TIMEOUT_MS = 10_000;
-
-/** Max time to wait for the worker to exit cooperatively on shutdown before
- *  force-terminating it. The worker drops queued work immediately but must
- *  finish any in-flight, uninterruptible single-threaded ONNX inference batch
- *  before it can exit — without this cap that could block process exit. */
-const WORKER_SHUTDOWN_TIMEOUT_MS = 1_500;
-
-/** Cooperative cancellation controls for bounded embedding phases.
- * `deadlineMs` is a duration from invocation, matching the existing shutdown
- * deadline convention in this module. */
-export interface EmbeddingOperationOptions {
-  signal?: AbortSignal;
-  deadlineMs?: number;
-}
-
-/** Stable phase names for orchestration callers (notably semantic lint). */
-export type EmbeddingAbortPhase =
-  | "provider-readiness"
-  | "settle-document-embeds"
-  | "knowledge-backfill";
-
-export type EmbeddingAbortCode = "aborted" | "deadline-exceeded";
-
-/** Typed cooperative-stop signal for embedding work. */
-export class EmbeddingAbortError extends Error {
-  readonly code: EmbeddingAbortCode;
-  readonly phase: EmbeddingAbortPhase;
-
-  constructor(
-    phase: EmbeddingAbortPhase,
-    code: EmbeddingAbortCode,
-    cause?: unknown,
-  ) {
-    super(
-      code === "deadline-exceeded"
-        ? `Embedding phase '${phase}' exceeded its deadline`
-        : `Embedding phase '${phase}' was aborted`,
-    );
-    this.name = "EmbeddingAbortError";
-    this.code = code;
-    this.phase = phase;
-    if (cause !== undefined)
-      (this as Error & { cause?: unknown }).cause = cause;
-  }
-}
-
-interface EmbeddingAbortGuard {
-  phase: EmbeddingAbortPhase;
-  signal?: AbortSignal;
-  deadlineAt?: number;
-}
-
-function abortCodeForReason(reason: unknown): EmbeddingAbortCode {
-  if (reason instanceof EmbeddingAbortError) return reason.code;
-  if (
-    typeof reason === "object" &&
-    reason !== null &&
-    "name" in reason &&
-    reason.name === "TimeoutError"
-  ) {
-    return "deadline-exceeded";
-  }
-  return "aborted";
-}
-
-function createEmbeddingAbortGuard(
-  phase: EmbeddingAbortPhase,
-  options: EmbeddingOperationOptions,
-): EmbeddingAbortGuard {
-  const deadlineMs = options.deadlineMs;
-  if (
-    deadlineMs !== undefined &&
-    (!Number.isFinite(deadlineMs) || deadlineMs < 0)
-  ) {
-    throw new RangeError(
-      "embedding deadlineMs must be a finite non-negative number",
-    );
-  }
-  return {
-    phase,
-    signal: options.signal,
-    deadlineAt: deadlineMs === undefined ? undefined : Date.now() + deadlineMs,
-  };
-}
-
-function currentEmbeddingAbort(
-  guard: EmbeddingAbortGuard,
-): EmbeddingAbortError | null {
-  if (guard.signal?.aborted) {
-    return new EmbeddingAbortError(
-      guard.phase,
-      abortCodeForReason(guard.signal.reason),
-      guard.signal.reason,
-    );
-  }
-  if (guard.deadlineAt !== undefined && Date.now() >= guard.deadlineAt) {
-    return new EmbeddingAbortError(guard.phase, "deadline-exceeded");
-  }
-  return null;
-}
-
-function throwIfEmbeddingAborted(guard: EmbeddingAbortGuard): void {
-  const error = currentEmbeddingAbort(guard);
-  if (error) throw error;
-}
-
-/** Race already-scheduled work against the caller's cancellation boundary.
- * The underlying inference is not forcibly interrupted, but this caller
- * returns promptly and its loop never schedules another batch. */
-async function awaitEmbeddingOperation<T>(
-  work: Promise<T>,
-  guard: EmbeddingAbortGuard,
-): Promise<T> {
-  const alreadyAborted = currentEmbeddingAbort(guard);
-  if (alreadyAborted) {
-    // The caller created `work` immediately before this check. Observe any later
-    // rejection even though orchestration is stopping, avoiding an unhandled
-    // worker/API rejection after the lint deadline has already been reported.
-    void work.catch(() => {});
-    throw alreadyAborted;
-  }
-  if (!guard.signal && guard.deadlineAt === undefined) return await work;
-
-  let timer: ReturnType<typeof setTimeout> | undefined;
-  let onAbort: (() => void) | undefined;
-  const aborted = new Promise<never>((_resolve, reject) => {
-    if (guard.signal) {
-      onAbort = () => {
-        reject(
-          new EmbeddingAbortError(
-            guard.phase,
-            abortCodeForReason(guard.signal?.reason),
-            guard.signal?.reason,
-          ),
-        );
-      };
-      guard.signal.addEventListener("abort", onAbort, { once: true });
-    }
-    if (guard.deadlineAt !== undefined) {
-      timer = setTimeout(
-        () => {
-          reject(new EmbeddingAbortError(guard.phase, "deadline-exceeded"));
-        },
-        Math.max(0, guard.deadlineAt - Date.now()),
-      );
-    }
-  });
-
-  try {
-    return await Promise.race([work, aborted]);
-  } finally {
-    if (timer) clearTimeout(timer);
-    if (onAbort && guard.signal) {
-      guard.signal.removeEventListener("abort", onAbort);
-    }
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Adaptive local-embedding token cap — persistence
-// ---------------------------------------------------------------------------
-// Pure cap math (memory model, clamp, backoff, reconcile) lives in
-// embedding-cap.ts so it can be unit-tested in isolation. Here we keep only the
-// db-backed persistence and the worker-lifecycle glue.
-
-/** kv_meta key for the persisted learned cap + the free memory at learn time. */
-const EMBED_CAP_KV_KEY = "lore:embedding_cap";
-
-/** Minimum interval between upward cap re-probe checks. Throttles the freemem
- *  read + comparison so embed() stays cheap; the cap only needs to be right
- *  when work is actively flowing, so checking on embed() is sufficient. */
-const EMBED_REPROBE_INTERVAL_MS = 5 * 60_000;
-
-function readPersistedEmbedCap(): PersistedEmbedCap | null {
-  try {
-    const row = db()
-      .query("SELECT value FROM kv_meta WHERE key = ?")
-      .get(EMBED_CAP_KV_KEY) as { value: string } | null;
-    if (!row) return null;
-    const parsed = JSON.parse(row.value) as Partial<PersistedEmbedCap>;
-    if (
-      typeof parsed.cap !== "number" ||
-      typeof parsed.freeMemBytes !== "number"
-    ) {
-      return null;
-    }
-    return {
-      cap: parsed.cap,
-      freeMemBytes: parsed.freeMemBytes,
-      ...(typeof parsed.knownBadCap === "number" && parsed.knownBadCap > 0
-        ? { knownBadCap: parsed.knownBadCap }
-        : {}),
-    };
-  } catch {
-    return null;
-  }
-}
-
-/** Test seam: override `process.constrainedMemory()` for container-aware sizing
- *  tests (null → use the real value). */
-let testConstrainedMemoryBytes: number | null = null;
-export function _setConstrainedMemoryForTest(bytes: number | null): void {
-  testConstrainedMemoryBytes = bytes;
-}
-
-/** Test seam: override the host free-memory read used by every cap-sizing
- *  decision ({@link containerFreeBytes}), so the pool-aware divisor and the
- *  per-request current-free clamp can be driven deterministically without
- *  depending on the CI box's actual RAM (null → real `os.freemem()`). */
-let testHostFreememBytes: number | null = null;
-export function _setContainerFreeForTest(bytes: number | null): void {
-  testHostFreememBytes = bytes;
-}
-
-/** The process's cgroup memory LIMIT in bytes (not free-within-limit), or `0` if
- *  unconstrained / unknown / unsupported by the runtime. `process.constrainedMemory()`
- *  is libuv-backed (cgroup v1 + v2, no hard-coded paths) and returns `0` when
- *  unconstrained; it is present in both Node (≥18.15) and Bun. Read live: it's
- *  cheap and the consuming paths are cap init/re-probe (infrequent) plus the pool
- *  growth gate (only while all workers are busy and below the ceiling — never in
- *  the single-worker or at-ceiling steady state). */
-function constrainedMemoryLimit(): number {
-  if (testConstrainedMemoryBytes != null) return testConstrainedMemoryBytes;
-  const fn = (process as { constrainedMemory?: () => number })
-    .constrainedMemory;
-  const v = typeof fn === "function" ? fn() : 0;
-  return Number.isFinite(v) && v > 0 ? v : 0;
-}
-
-/** Container-aware live free memory: `os.freemem()` clamped to the cgroup memory
- *  limit ({@link clampFreeToContainerLimit}) so no sizing decision trusts
- *  host-level free memory inside a memory-capped container. A no-op (returns
- *  `freemem()` unchanged) when unconstrained or when the container limit exceeds
- *  host-reported free — so behavior on non-containerized / roomy hosts is
- *  identical to the pre-cgroup path. Use this for every memory-SIZING read;
- *  telemetry keeps raw `freemem()` so host vs container stays diagnosable.
- *
- *  Note the clamp uses the cgroup LIMIT, not free-within-limit: `min(hostFree,
- *  limit)` can slightly overestimate available memory in a mid-size container
- *  already using part of its budget, but that's deliberate — it's strictly ≤ the
- *  pre-fix host figure, and the per-worker budget already reserves the transient
- *  attention peak as headroom. Using cgroup free (`availableMemory()`) instead
- *  would break the no-op guarantee on constrained-but-roomy hosts. */
-function containerFreeBytes(): number {
-  const raw = testHostFreememBytes != null ? testHostFreememBytes : freemem();
-  return clampFreeToContainerLimit(raw, constrainedMemoryLimit());
-}
-
-function persistEmbedCap(
-  cap: number,
-  freeMemBytes: number = containerFreeBytes(),
-  knownBadCap = 0,
-): void {
-  try {
-    const value = JSON.stringify({
-      cap,
-      freeMemBytes,
-      ...(knownBadCap > 0 ? { knownBadCap } : {}),
-    } satisfies PersistedEmbedCap);
-    db()
-      .query(
-        "INSERT INTO kv_meta (key, value) VALUES (?, ?) ON CONFLICT(key) DO UPDATE SET value = ?",
-      )
-      .run(EMBED_CAP_KV_KEY, value, value);
-  } catch {
-    // Best-effort: a failure just means we re-derive the cap next start.
-  }
-}
-
-/**
- * Compute the token cap to (re)start a worker with. Prefers a learned,
- * persisted cap when current free memory is close to what it was learned at —
- * this avoids re-walking the backoff on every process restart (the "recurs on
- * every restart" failure). Otherwise it sizes from the memory model and
- * current free memory; a materially larger free pool lets the cap re-probe
- * upward on restart (in lieu of continuous additive increase). Always clamped.
- *
- * `memDivisor` is the number of pool workers this cap must share free memory
- * with (the pool ceiling; 1 for a standalone provider). Each worker sizes its
- * transient O(L²) attention budget from `free / memDivisor`, so `memDivisor`
- * workers running concurrently collectively stay within one worker's worth of
- * the `EMBED_MEM_FRACTION` budget — otherwise N independently-sized workers each
- * claim half of free memory and their sum OOMs the host (the exact native
- * SIGKILL the WASM heap wall used to prevent; see EmbeddingPool).
- */
-function computeInitialEmbedCap(
-  persisted: PersistedEmbedCap | null = readPersistedEmbedCap(),
-  memDivisor = 1,
-): number {
-  const free = containerFreeBytes() / Math.max(1, memDivisor);
-  return reconcileEmbedCap(
-    free,
-    persisted,
-    memoryModelEmbedCap(free),
-    persisted?.knownBadCap ?? 0,
-  );
-}
-
-/**
- * For tests: persist a learned embedding cap (kv_meta round-trip).
- *
- * `freeMemBytes` defaults to the live `freemem()`. Tests that need a
- * deterministic start cap pass `0`: `reconcileEmbedCap` treats a non-positive
- * learn-time baseline as ratio = 1, so the persisted cap is always trusted
- * as-is — independent of how much free memory the host happens to have. Without
- * this, a host whose free memory swings >25% between persist and provider
- * construction (e.g. a CI box running the full suite in parallel) reconciles to
- * the freemem-derived model cap instead, making the start cap non-deterministic.
- */
-export function _persistEmbedCap(
-  cap: number,
-  freeMemBytes?: number,
-  knownBadCap?: number,
-): void {
-  persistEmbedCap(cap, freeMemBytes, knownBadCap);
-}
-
-/** For tests: read the persisted embedding cap (or null when absent/corrupt). */
-export function _readPersistedEmbedCap(): PersistedEmbedCap | null {
-  return readPersistedEmbedCap();
-}
-
-/**
- * Coarse per-text character bound applied before a text is cloned across the
- * worker boundary — only a payload-size guard. The real, memory-aware token
- * cap is applied inside the worker (`WorkerInitData.maxTokens`). Sized at the
- * model's max sequence length so it never caps below what a healthy host could
- * embed.
- */
-const LOCAL_MAX_CHARS = MODEL_MAX_TOKENS * 4; // ~8192 tokens × ~4 chars/token
-
-/**
- * Truncate a string to LOCAL_MAX_CHARS without splitting a UTF-16 surrogate pair.
- * If the cut falls on a high surrogate (0xD800-0xDBFF), backs up one char.
- */
-function safeLocalTruncate(text: string): string {
-  if (text.length <= LOCAL_MAX_CHARS) return text;
-  let end = LOCAL_MAX_CHARS;
-  const code = text.charCodeAt(end - 1);
-  if (code >= 0xd800 && code <= 0xdbff) end--; // don't split surrogate pair
-  return text.slice(0, end);
-}
-
-// ---------------------------------------------------------------------------
-// Embedding failure telemetry hook (wired by the gateway to Sentry)
-// ---------------------------------------------------------------------------
-
-/** Context for an embedding-worker memory failure, surfaced to an optional host
- *  telemetry hook. @loreai/core stays Sentry-free; the gateway wires this up. */
-export interface EmbeddingFailureInfo {
-  /** "oom-backoff" = recovered by lowering the cap + respawning;
-   *  "floor-latch" = hit MIN_EMBED_TOKENS and degraded to FTS-only. */
-  kind: "oom-backoff" | "floor-latch";
-  capBefore: number;
-  capAfter: number;
-  /** Number of in-flight requests at the time of the OOM. */
-  batchSize: number;
-  /** Longest input text (chars) among in-flight requests. */
-  longestChars: number;
-  freeMemBytes: number;
-  rssBytes: number;
-}
-
-let embeddingFailureHook: ((info: EmbeddingFailureInfo) => void) | null = null;
-
-/** Register a host telemetry hook fired on embedding-worker OOM backoff/latch.
- *  Pass null to clear. The hook must not throw; errors are swallowed. */
-export function setEmbeddingFailureHook(
-  fn: ((info: EmbeddingFailureInfo) => void) | null,
-): void {
-  embeddingFailureHook = fn;
-}
-
-function fireEmbeddingFailure(
-  info: Omit<EmbeddingFailureInfo, "freeMemBytes" | "rssBytes">,
-): void {
-  const hook = embeddingFailureHook;
-  if (!hook) return;
-  try {
-    hook({
-      ...info,
-      freeMemBytes: freemem(),
-      rssBytes: process.memoryUsage().rss,
-    });
-  } catch {
-    // Telemetry must never break the embedding path.
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Provider interface
-// ---------------------------------------------------------------------------
-
-export interface EmbeddingProvider {
-  embed(
-    texts: string[],
-    inputType: "document" | "query",
-    signal?: AbortSignal,
-  ): Promise<Float32Array[]>;
-  readonly maxBatchSize: number;
-}
-
-// ---------------------------------------------------------------------------
-// Voyage AI provider
-// ---------------------------------------------------------------------------
-
-const VOYAGE_API_URL = "https://api.voyageai.com/v1/embeddings";
-
-type VoyageResponse = {
-  data: Array<{ embedding: number[]; index: number }>;
-  model: string;
-  usage: { total_tokens: number };
-};
-
-class VoyageProvider implements EmbeddingProvider {
-  readonly maxBatchSize = 128;
-  private apiKey: string;
-  private model: string;
-  private dimensions: number;
-
-  constructor(apiKey: string, model: string, dimensions: number) {
-    this.apiKey = apiKey;
-    this.model = model;
-    this.dimensions = dimensions;
-  }
-
-  async embed(
-    texts: string[],
-    inputType: "document" | "query",
-    signal?: AbortSignal,
-  ): Promise<Float32Array[]> {
-    let res: Response;
-    try {
-      res = await fetch(VOYAGE_API_URL, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${this.apiKey}`,
-        },
-        body: JSON.stringify({
-          input: texts,
-          model: this.model,
-          input_type: inputType,
-          output_dimension: this.dimensions,
-        }),
-        signal: signal
-          ? AbortSignal.any([signal, AbortSignal.timeout(EMBED_TIMEOUT_MS)])
-          : AbortSignal.timeout(EMBED_TIMEOUT_MS),
-      });
-    } catch {
-      throw new Error("Voyage embeddings API request failed");
-    }
-
-    if (!res.ok) {
-      throw new Error(`Voyage embeddings API failed with HTTP ${res.status}`);
-    }
-
-    let json: VoyageResponse;
-    try {
-      json = (await res.json()) as VoyageResponse;
-    } catch {
-      throw new Error("Voyage embeddings API returned malformed JSON");
-    }
-    const sorted = [...json.data].sort((a, b) => a.index - b.index);
-    return sorted.map((d) => new Float32Array(d.embedding));
-  }
-}
-
-// ---------------------------------------------------------------------------
-// OpenAI provider
-// ---------------------------------------------------------------------------
-
-const OPENAI_API_URL = "https://api.openai.com/v1/embeddings";
-
-type OpenAIResponse = {
-  data: Array<{ embedding: number[]; index: number }>;
-  model: string;
-  usage: { prompt_tokens: number; total_tokens: number };
-};
-
-class OpenAIProvider implements EmbeddingProvider {
-  readonly maxBatchSize = 2048;
-  private apiKey: string;
-  private model: string;
-  private dimensions: number;
-
-  constructor(apiKey: string, model: string, dimensions: number) {
-    this.apiKey = apiKey;
-    this.model = model;
-    this.dimensions = dimensions;
-  }
-
-  async embed(
-    texts: string[],
-    _inputType: "document" | "query",
-    signal?: AbortSignal,
-  ): Promise<Float32Array[]> {
-    const body: Record<string, unknown> = {
-      input: texts,
-      model: this.model,
-    };
-    // OpenAI supports dimensions parameter for text-embedding-3-* models
-    if (this.model.startsWith("text-embedding-3")) {
-      body.dimensions = this.dimensions;
-    }
-
-    const res = await fetch(OPENAI_API_URL, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${this.apiKey}`,
-      },
-      body: JSON.stringify(body),
-      signal: signal
-        ? AbortSignal.any([signal, AbortSignal.timeout(EMBED_TIMEOUT_MS)])
-        : AbortSignal.timeout(EMBED_TIMEOUT_MS),
-    });
-
-    if (!res.ok) {
-      throw new Error(`OpenAI embeddings API failed with HTTP ${res.status}`);
-    }
-
-    const json = (await res.json()) as OpenAIResponse;
-    const sorted = [...json.data].sort((a, b) => a.index - b.index);
-    return sorted.map((d) => new Float32Array(d.embedding));
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Local provider (@huggingface/transformers + nomic-embed-text-v1.5)
-// ---------------------------------------------------------------------------
-
-/**
- * Thrown when `LocalProvider` cannot initialize (e.g. ONNX runtime fails
- * to load). Callers in `recall.ts` / `ltm.ts` / `distillation.ts` gate
- * on `isAvailable()`, which flips to `false` after this error fires once.
- */
-export class LocalProviderUnavailableError extends Error {
-  constructor(cause?: unknown) {
-    super(
-      "Local embedding provider unavailable: '@huggingface/transformers' failed to initialize. " +
-        "Recall will use FTS-only search. To use a remote provider instead, set " +
-        "search.embeddings.provider to 'voyage' or 'openai' in .lore.json " +
-        "and provide the corresponding API key (VOYAGE_API_KEY / OPENAI_API_KEY).",
-    );
-    this.name = "LocalProviderUnavailableError";
-    if (cause !== undefined)
-      (this as Error & { cause?: unknown }).cause = cause;
-  }
-}
-
-/** A caller stopped waiting for an embedding request. */
-export class EmbeddingRequestAbortedError extends Error {
-  constructor() {
-    super("Embedding request aborted");
-    this.name = "EmbeddingRequestAbortedError";
-  }
-}
-
-/** The host-owned local queue hit its bounded count or byte budget. The
- * provider remains healthy; durable callers should retry with backoff. */
-export class EmbeddingQueueCapacityError extends Error {
-  constructor() {
-    super("Embedding queue capacity exceeded");
-    this.name = "EmbeddingQueueCapacityError";
-  }
-}
-
-/** The worker exceeded its own initialization or execution lifetime. This is
- * independent of every caller's deadline and retires only the affected slot. */
-export class EmbeddingWorkerWatchdogError extends Error {
-  readonly stage: "init" | "execution";
-
-  constructor(stage: "init" | "execution") {
-    super(`Embedding worker ${stage} watchdog expired`);
-    this.name = "EmbeddingWorkerWatchdogError";
-    this.stage = stage;
-  }
-}
-
-type LocalProviderFailureCause = "terminal" | "transient-init-exhausted";
-/** A terminal cause must dominate concurrent transient failures. Otherwise a
- * late healthy sibling could clear a missing-stack / WASM / OOM latch after
- * another slot exhausts the transient retry budget. */
-let localProviderFailureCause: LocalProviderFailureCause | null = null;
-let localProviderErrorLogged = false;
-/** True when the latch cause is the ONE that never recovers on a re-probe: the
- *  optional embedding stack is absent (needs a reinstall, not a retry). Every
- *  other latch cause (WASM-fatal, floor OOM, exhausted transient-init budget, a
- *  non-zero worker exit) can plausibly clear on a fresh worker, so self-heal
- *  re-probes those on a slow cadence. */
-let localStackMissing = false;
-/** Epoch ms of the last self-heal re-probe; `0` = never (primed on first tick). */
-let lastSelfHealAt = 0;
-/** How long to wait between self-heal re-probes of a latched local provider. */
-const SELF_HEAL_INTERVAL_MS = 6 * 60 * 60 * 1000; // 6h
-
-// --- Transient init-failure retry (self-heal a one-off worker init failure) ---
-// A single model-init failure used to latch the provider FTS-only for the whole
-// process lifetime. But a transient failure — e.g. the ONNX model file read
-// while a concurrent process was still writing it during a multi-instance
-// restart, or momentary memory pressure — should self-heal: reject the in-flight
-// requests, wait a cooldown, then respawn a FRESH worker (a new init attempt) on
-// the next embed. Only latch permanently after LOCAL_INIT_MAX_ATTEMPTS
-// consecutive failures, or immediately for a cause that will never self-heal
-// (the optional stack being absent).
-const LOCAL_INIT_MAX_ATTEMPTS = 3;
-let localInitFailures = 0;
-/** Monotonic generation for transient failures. Recovery slots capture this at
- * admission and may clear retry debt only if no newer failure has occurred. */
-let localInitFailureGeneration = 0;
-/** Epoch ms; `> 0` means "cooling down after a transient init failure — retry a
- *  fresh worker once `Date.now()` reaches it". Reset to 0 on retry/recovery. */
-let localInitRetryAt = 0;
-// Ceiling for the init-retry cooldown; the first retries are much faster (see
-// computeInitRetryDelayMs). Also the test seam: 0 → retry immediately.
-let localInitCooldownMs = 30_000;
-
-function latchLocalProvider(cause: LocalProviderFailureCause): void {
-  if (cause === "terminal" || localProviderFailureCause === null) {
-    localProviderFailureCause = cause;
-  }
-}
-
-function clearLocalProviderLatch(): void {
-  localProviderFailureCause = null;
-}
-
-/** Base delay for the first init-retry, and the geometric backoff factor. A
- *  transient init failure (e.g. a cold ORT "protobuf parsing failed" in the Bun
- *  worker) reliably recovers on the NEXT fresh-worker respawn, so the first
- *  retry should be fast; only a genuine failure (which fails the fast retry
- *  too) needs to back off toward the ceiling to avoid a respawn storm. With
- *  LOCAL_INIT_MAX_ATTEMPTS=3 the two retry waits are 2s then 8s. */
-const INIT_RETRY_BASE_MS = 2_000;
-const INIT_RETRY_BACKOFF_FACTOR = 4;
-
-/** Cooldown before the Nth fresh-worker init-retry: `base * factor^(N-1)`,
- *  clamped to `ceilingMs`. Pure (no module state) so the schedule is unit
- *  testable. `ceilingMs=0` (the test seam) collapses to 0 → retry immediately. */
-export function computeInitRetryDelayMs(
-  failures: number,
-  ceilingMs: number,
-): number {
-  const n = Math.max(1, failures);
-  const backoff = INIT_RETRY_BASE_MS * INIT_RETRY_BACKOFF_FACTOR ** (n - 1);
-  return Math.min(ceilingMs, backoff);
-}
-
-/** Test seam: shorten the init-retry cooldown ceiling so the retry path is
- *  drivable without real time (`0` → the next availability check retries
- *  immediately; `null` restores the production default). */
-export function _setLocalInitCooldownMsForTest(ms: number | null): void {
-  localInitCooldownMs = ms ?? 30_000;
-}
-
-/** For tests: observe the pending init-retry deadline (epoch ms; 0 = none). */
-export function _getLocalInitRetryAtForTest(): number {
-  return localInitRetryAt;
-}
-
-/** For tests: arm/clear the init-retry deadline (epoch ms; 0 = none) so the
- *  transient "retrying" health state is drivable without a fake worker. */
-export function _setLocalInitRetryAtForTest(at: number): void {
-  localInitRetryAt = at;
-}
-
-/** For tests: reset the local provider probe + transient-retry state. */
-export function _resetLocalProviderProbe(): void {
-  clearLocalProviderLatch();
-  localProviderErrorLogged = false;
-  localStackMissing = false;
-  localInitFailures = 0;
-  localInitFailureGeneration = 0;
-  localInitRetryAt = 0;
-  lastSelfHealAt = 0;
-}
-
-/** For tests: simulate the local provider being unavailable, without
- *  actually spawning a worker. After this call, `isAvailable()` returns
- *  false for the local provider. Pass `stackMissing: true` to simulate the
- *  ONE terminal (non-self-healing) cause. */
-export function _markLocalProviderUnavailable(stackMissing = false): void {
-  latchLocalProvider("terminal");
-  localStackMissing = stackMissing;
-  localProviderErrorLogged = true; // suppress the info log in tests
-}
-
-/**
- * Self-heal a permanently-latched LOCAL embedding provider: on a slow cadence,
- * clear the latch so the next embed spawns a fresh worker and re-attempts init.
- * Only recoverable causes are re-probed — a missing optional stack
- * (`localStackMissing`) is skipped, since only a reinstall fixes that. Safe to
- * call every idle tick: it self-gates on the latch, the cause, and
- * `SELF_HEAL_INTERVAL_MS`, and primes its clock on the first call so it never
- * re-probes immediately after a fresh latch. Returns true iff it cleared the
- * latch on this call. `nowMs` is injectable for tests.
- */
-export function maybeSelfHealEmbeddingProvider(nowMs = Date.now()): boolean {
-  if (localProviderFailureCause === null || localStackMissing) return false;
-  if (lastSelfHealAt === 0) {
-    // Prime: wait a full interval before the first re-probe.
-    lastSelfHealAt = nowMs;
-    return false;
-  }
-  if (nowMs - lastSelfHealAt < SELF_HEAL_INTERVAL_MS) return false;
-  lastSelfHealAt = nowMs;
-  // Clear the latch + transient counters so getProvider() rebuilds a fresh pool
-  // and the next embed re-inits. If it fails again it simply re-latches — cheap
-  // at this cadence.
-  clearLocalProviderLatch();
-  localProviderErrorLogged = false;
-  localInitFailures = 0;
-  localInitFailureGeneration = 0;
-  localInitRetryAt = 0;
-  log.info(
-    "self-heal: re-probing a previously-latched local embedding provider " +
-      "(a fresh worker retries init on the next embed; re-latches if it fails again)",
-  );
-  void resetProvider().catch((error: unknown) => {
-    log.error("self-heal embedding worker shutdown was not confirmed", error);
-  });
-  return true;
-}
-
-type EmbeddingWorkerSpawnOptions =
-  import("node:worker_threads").WorkerOptions & {
-    /** Virtual filename used by the SEA eval worker's CJS require shim. */
-    filename?: string;
-  };
-
-/** Test seam: when set, `ensureWorker()` uses this factory instead of spawning
- *  a real `node:worker_threads` Worker, so worker lifecycle and construction
- *  options can be driven deterministically. Never set in production. */
-let testWorkerFactory:
-  | ((
-      data: WorkerInitData,
-      entrypoint: string | URL,
-      options: EmbeddingWorkerSpawnOptions,
-    ) => import("node:worker_threads").Worker)
-  | null = null;
-
-/** For tests: install the worker factory seam above (null clears it). */
-export function _setTestWorkerFactory(
-  factory:
-    | ((
-        data: WorkerInitData,
-        entrypoint: string | URL,
-        options: EmbeddingWorkerSpawnOptions,
-      ) => import("node:worker_threads").Worker)
-    | null,
-): void {
-  testWorkerFactory = factory;
-}
-
-const WORKER_DIAGNOSTIC_MAX_CHARS = 2_000;
-const WORKER_DIAGNOSTIC_WINDOW_MS = 60_000;
-const WORKER_DIAGNOSTIC_LINES_PER_WINDOW = 20;
-
-/** Continuously consume an owned worker stream. Diagnostics are bounded and
- * routed through the parent logger (which honors TUI stderr silencing and never
- * writes stdout) rather than piped to either parent stdio stream. The rate cap
- * keeps a noisy native dependency from turning output draining into synchronous
- * log-file backpressure; excess bytes are still consumed immediately. */
-function drainEmbeddingWorkerStream(
-  stream: import("node:stream").Readable | null | undefined,
-  source: "stdout" | "stderr",
-): void {
-  if (!stream) return; // Test doubles created before the owned-stdio contract.
-
-  let buffered = "";
-  let emitted = 0;
-  let suppressed = 0;
-  let windowStartedAt = Date.now();
-  let finalized = false;
-
-  const logDiagnostic = (message: string): void => {
-    try {
-      log.info(message);
-    } catch {
-      // A host logger/sink must never interrupt stream consumption and let the
-      // worker block on a full pipe. Structured worker messages still carry all
-      // actionable failures to the normal parent handlers.
-    }
-  };
-
-  const route = (rawLine: string): void => {
-    let line = "";
-    const end = rawLine.endsWith("\r") ? rawLine.length - 1 : rawLine.length;
-    for (let i = 0; i < end && line.length < WORKER_DIAGNOSTIC_MAX_CHARS; i++) {
-      const code = rawLine.charCodeAt(i);
-      if (
-        code <= 0x08 ||
-        code === 0x0b ||
-        code === 0x0c ||
-        (code >= 0x0e && code <= 0x1f) ||
-        code === 0x7f
-      ) {
-        continue;
-      }
-      line += rawLine[i];
-    }
-    if (!line) return;
-
-    const now = Date.now();
-    if (now - windowStartedAt >= WORKER_DIAGNOSTIC_WINDOW_MS) {
-      if (suppressed > 0) {
-        logDiagnostic(
-          `embedding worker ${source}: suppressed ${suppressed} noisy diagnostic line(s)`,
-        );
-      }
-      windowStartedAt = now;
-      emitted = 0;
-      suppressed = 0;
-    }
-
-    if (emitted < WORKER_DIAGNOSTIC_LINES_PER_WINDOW) {
-      emitted++;
-      logDiagnostic(`embedding worker ${source}: ${line}`);
-    } else {
-      suppressed++;
-    }
-  };
-
-  const consume = (chunk: string): void => {
-    buffered += chunk;
-    for (;;) {
-      const newline = buffered.indexOf("\n");
-      if (newline >= 0) {
-        route(buffered.slice(0, newline));
-        buffered = buffered.slice(newline + 1);
-        continue;
-      }
-      if (buffered.length > WORKER_DIAGNOSTIC_MAX_CHARS) {
-        route(`${buffered.slice(0, WORKER_DIAGNOSTIC_MAX_CHARS)}…`);
-        buffered = buffered.slice(WORKER_DIAGNOSTIC_MAX_CHARS);
-        continue;
-      }
-      break;
-    }
-  };
-
-  const finalize = (): void => {
-    if (finalized) return;
-    finalized = true;
-    if (buffered) route(buffered);
-    buffered = "";
-    if (suppressed > 0) {
-      logDiagnostic(
-        `embedding worker ${source}: suppressed ${suppressed} noisy diagnostic line(s)`,
-      );
-    }
-  };
-
-  stream.setEncoding("utf8");
-  stream.on("data", consume);
-  stream.on("end", finalize);
-  stream.on("close", finalize);
-  stream.on("error", (error) => {
-    route(`diagnostic stream error: ${error.message}`);
-    finalize();
-  });
-  stream.resume();
-}
-
-function drainEmbeddingWorkerOutput(
-  worker: import("node:worker_threads").Worker,
-): void {
-  drainEmbeddingWorkerStream(worker.stdout, "stdout");
-  drainEmbeddingWorkerStream(worker.stderr, "stderr");
-}
-
-/** True iff the local provider has been probed and found broken. */
-function localProviderKnownUnavailable(): boolean {
-  return localProviderFailureCause !== null;
-}
-
-/**
- * Local embedding provider using @huggingface/transformers with
- * nomic-embed-text-v1.5 by default.
- *
- * No API key required — runs entirely on-device via ONNX Runtime.
- * Model files are downloaded on first use (~137MB for INT8 quantized)
- * and cached locally. Subsequent inits load from cache.
- *
- * ONNX inference runs in a dedicated `node:worker_threads` Worker so the
- * main thread's event loop stays free. This class is a thin RPC client —
- * it posts `{ texts, inputType }` to the worker and awaits a reply.
- * The worker owns the transformers.js pipeline and processes requests
- * sequentially from a priority queue (recall queries jump ahead of
- * backfill batches).
- *
- * Task instruction prefixes are prepended automatically:
- *   - "document" → "search_document: <text>"
- *   - "query"    → "search_query: <text>"
- */
-class LocalProvider implements EmbeddingProvider {
-  // With inference off the main thread, large batches no longer block
-  // the event loop. 256 maximises throughput per round-trip to the
-  // worker. Backfill callers use token-budget-based batching (see
-  // nextBatch) to give the worker's priority queue breathing room
-  // for recall queries and prevent OOM on long texts.
-  readonly maxBatchSize = 256;
-
-  private worker: import("node:worker_threads").Worker | null = null;
-  private workerReady = false;
-  private workerInitError: string | null = null;
-  private pendingRequests = new Map<
-    number,
-    {
-      resolve: (vectors: Float32Array[]) => void;
-      reject: (error: Error) => void;
-      /** Original request payload, retained so it can be re-submitted to a
-       *  freshly respawned worker after an OOM backoff (fresh WASM heap). */
-      payload: EmbedRequest;
-      onExecutionStart?: () => void;
-    }
-  >();
-  private nextRequestId = 0;
-  private initPromise: Promise<void> | null = null;
-  private closing = false;
-  private shutdownPromise: Promise<void> | null = null;
-  /** Superseded workers remain owned until terminate() confirms their exit. */
-  private readonly retiredWorkerShutdowns = new Set<Promise<void>>();
-  /** A rejected termination remains fatal for this provider generation. */
-  private readonly retiredWorkerShutdownErrors: unknown[] = [];
-  private modelId: string;
-  private dimensions: number;
-  /** Memory-aware input token cap, owned by the main thread and passed to the
-   *  worker via workerData. Lowered ×0.7 per OOM respawn; persisted so restarts
-   *  start at the learned value instead of re-walking the backoff. */
-  private maxTokens: number;
-  /** Free memory (bytes) when `maxTokens` was last (re)learned. The upward
-   *  re-probe only fires when current free memory exceeds this by
-   *  EMBED_REPROBE_RATIO — i.e. when memory has genuinely recovered, which
-   *  distinguishes transient starvation from a real hardware limit. */
-  private capFreememAtLearn: number;
-  /** Timestamp (ms) of the last upward re-probe check, for throttling. */
-  private lastReprobeAt = 0;
-  /** Highest cap (tokens) that has OOMed in this process, or 0 if none. The
-   *  upward re-probe never climbs to or past it — a rising os.freemem() does not
-   *  prove the WASM heap can grow that far. Cleared only by a process restart. */
-  private lastOomCap = 0;
-  /** Number of pool workers this provider shares free memory with (the pool
-   *  ceiling; 1 for a standalone provider). Every free-memory-derived cap is
-   *  sized from `free / memDivisor` so N concurrent workers can't each claim a
-   *  full `EMBED_MEM_FRACTION` share and sum past the host's RAM (BUG: native
-   *  OOM is an uncatchable SIGKILL, so this must be prevented, not recovered). */
-  private readonly memDivisor: number;
-  /** Force the worker onto the bundled WASM ONNX Runtime, set after a native
-   *  worker reported it couldn't parse an intact model (`init-needs-wasm`, the
-   *  Bun ↔ onnxruntime-node incompatibility, #1379). Threaded into every
-   *  subsequent `workerData` so respawns (incl. OOM backoff) stay on WASM. */
-  private forceWasm = false;
-  /** One-shot latch: we respawn once onto WASM after `init-needs-wasm`. A second
-   *  `init-needs-wasm` (WASM also failing, which shouldn't happen) is treated as
-   *  a genuine init failure instead of looping respawns. */
-  private wasmFallbackTried = false;
-  private readonly onUnavailable?: () => void;
-
-  constructor(
-    modelId: string,
-    dimensions: number,
-    memDivisor = 1,
-    onUnavailable?: () => void,
-  ) {
-    this.modelId = modelId;
-    this.dimensions = dimensions;
-    this.memDivisor = Math.max(1, memDivisor);
-    this.onUnavailable = onUnavailable;
-    // Seed lastOomCap from the persisted known-bad cap so an upward re-probe in
-    // THIS process still respects a ceiling the WASM heap rejected in a PRIOR
-    // one (a rising freemem doesn't prove the fixed heap grew). Read once and
-    // reuse for the initial cap to avoid a second kv_meta round-trip.
-    const persisted = readPersistedEmbedCap();
-    this.lastOomCap = persisted?.knownBadCap ?? 0;
-    this.maxTokens = computeInitialEmbedCap(persisted, this.memDivisor);
-    this.capFreememAtLearn = containerFreeBytes();
-  }
-
-  /** The token cap to send with the NEXT request: the learned `maxTokens`
-   *  clamped to what CURRENT free memory supports for this worker's share of
-   *  the pool. Re-evaluated per request (not just at construction) because free
-   *  memory drops as concurrent sessions and sibling workers allocate — and on
-   *  the native ONNX path a too-large allocation triggers the OS OOM-killer
-   *  (uncatchable SIGKILL) instead of a recoverable in-worker OOM. Clamping down
-   *  to live free memory avoids the allocation the ×0.7 backoff can't catch. */
-  private effectiveMaxTokens(): number {
-    const liveCap = memoryModelEmbedCap(containerFreeBytes() / this.memDivisor);
-    return Math.min(this.maxTokens, liveCap);
-  }
-
-  /**
-   * Ensure the worker thread is running. Worker startup failure is
-   * surfaced as `LocalProviderUnavailableError` to mark the provider as
-   * broken and degrade to FTS-only search.
-   */
-  private async ensureWorker(): Promise<void> {
-    if (this.closing)
-      throw new LocalProviderUnavailableError("embedding worker is closing");
-    if (this.workerReady) return;
-    if (this.workerInitError)
-      throw new LocalProviderUnavailableError(this.workerInitError);
-    if (this.initPromise) return this.initPromise;
-
-    this.initPromise = (async () => {
-      // Fast-fail if a previous attempt already marked local broken.
-      if (localProviderFailureCause !== null) {
-        throw new LocalProviderUnavailableError();
-      }
-
-      const { Worker } = await import("node:worker_threads");
-      if (this.closing) {
-        throw new LocalProviderUnavailableError("embedding worker is closing");
-      }
-
-      // Resolve how to spawn the worker.
-      //
-      // In fossilize SEA binary mode: the binary's sea-entry.ts reads
-      // the worker source from the SEA asset and exposes it via
-      // `globalThis.__LORE_WORKER_SOURCE__`. We pass it to
-      // `new Worker(code, { eval: true, filename, workerData })`.
-      // The `filename` option sets `__filename` inside the worker to
-      // an absolute path, so the post-processing patch that replaces
-      // `createRequire(shim.url)` with
-      // `createRequire(pathToFileURL(__filename).href)` resolves
-      // correctly. No file is written to disk — the filename is
-      // purely virtual.
-      //
-      // In CJS bundles (gateway npm package) and dev: use the sibling
-      // embedding-worker.{cjs,js,ts} file as the worker entrypoint.
-      const workerSource = (globalThis as Record<string, unknown>)
-        .__LORE_WORKER_SOURCE__ as string | undefined;
-      const vendor = vendorModelInfo();
-      const workerInitData: WorkerInitData = {
-        modelId: this.modelId,
-        dimensions: this.dimensions,
-        // Only a fallback for a request that omits maxTokens (every embed()
-        // carries one) — but size it to current free memory too, so even that
-        // fallback path can't drive a native over-allocation.
-        maxTokens: this.effectiveMaxTokens(),
-        // Cgroup-CPU-aware native ORT intra-op cap, computed here on the main
-        // thread (the worker can't value-import ort-native). undefined = no-op
-        // (unconstrained host); applied on the native path only in the worker.
-        intraOpThreads: nativeIntraOpThreads(),
-        vendorModel: vendor ? { localModelPath: vendor.localModelPath } : null,
-        // Snapshot the host's silence state — the worker's own `globalThis`
-        // can't see the main thread's flag (re-read on every OOM respawn).
-        stderrSilenced: log.isStderrSilenced(),
-        // Force WASM once a native worker proved it couldn't parse the model
-        // (#1379). Sticky across respawns so we don't bounce back to native.
-        forceWasm: this.forceWasm,
-      };
-
-      let workerEntrypoint: string | URL;
-      let workerOptions: EmbeddingWorkerSpawnOptions;
-      if (workerSource !== undefined) {
-        const path = await import("node:path");
-        const os = await import("node:os");
-        workerEntrypoint = workerSource;
-        workerOptions = {
-          eval: true,
-          filename: path.join(os.homedir(), ".cache", "lore", "worker.cjs"),
-          workerData: workerInitData,
-          // Own both streams. Without these flags Node inherits parent stdio,
-          // so worker console output bypasses parent-only JSON/report capture.
-          stdout: true,
-          stderr: true,
-        };
-      } else {
-        // npm bundle / dev path: point at a sibling worker file.
-        // CJS uses __filename (always defined); ESM uses import.meta.url.
-        let workerUrl: string | URL;
-        if (typeof __filename === "string") {
-          const { pathToFileURL } = await import("node:url");
-          // Match the sibling worker file extension to the current bundle:
-          //   .ts  → dev (vitest/tsx)
-          //   .cjs → gateway CJS npm bundle
-          //   .js  → core ESM npm bundle (fallback)
-          const workerExt = __filename.endsWith(".ts")
-            ? ".ts"
-            : __filename.endsWith(".cjs")
-              ? ".cjs"
-              : ".js";
-          workerUrl = new URL(
-            `./embedding-worker${workerExt}`,
-            pathToFileURL(__filename),
-          );
-        } else {
-          // ESM (Bun, tsx): resolve worker relative to this module's URL.
-          // In CJS bundles the gateway build script (script/bundle.ts)
-          // rewrites `import.meta.url` to an injected `import_meta_url`
-          // shim — see packages/gateway/script/import-meta-url.js. This
-          // branch is unreachable in CJS at runtime since __filename is
-          // always defined there, but the shim keeps the source natural
-          // and silences esbuild's `empty-import-meta` static warning.
-          const selfUrl = import.meta.url;
-          workerUrl = new URL(
-            `./embedding-worker${selfUrl.endsWith(".ts") ? ".ts" : ".js"}`,
-            selfUrl,
-          );
-        }
-        workerEntrypoint = workerUrl;
-        workerOptions = {
-          workerData: workerInitData,
-          stdout: true,
-          stderr: true,
-        };
-      }
-
-      if (testWorkerFactory) {
-        // Test seam (never set in production): deterministic fake workers can
-        // inspect the exact options used by both file-backed and SEA branches.
-        this.worker = testWorkerFactory(
-          workerInitData,
-          workerEntrypoint,
-          workerOptions,
-        );
-      } else {
-        this.worker = new Worker(workerEntrypoint, workerOptions);
-      }
-
-      // Attach flowing readers before any request is posted. The streams remain
-      // owned and drained for the worker's whole lifetime, including init/OOM
-      // diagnostics emitted before the first response or during shutdown.
-      drainEmbeddingWorkerOutput(this.worker);
-
-      // Don't let the worker prevent process exit.
-      this.worker.unref();
-
-      // Capture the worker THIS init spawned. Every event handler below is bound
-      // to `spawned` and early-returns if `this.worker !== spawned` — a stale
-      // worker's late events (e.g. the `exit(1)` that `terminate()` emits during
-      // a WASM fallback respawn, #1379/#1387-B1) must never clobber the fresh
-      // worker's state, reject its resubmitted requests, or latch the provider.
-      const spawned = this.worker;
-
-      // Wire up response handler.
-      this.worker.on("message", (msg: WorkerOutbound) => {
-        if (this.worker !== spawned) return; // superseded worker — ignore
-        switch (msg.type) {
-          case "started": {
-            this.pendingRequests.get(msg.id)?.onExecutionStart?.();
-            break;
-          }
-          case "result": {
-            const pending = this.pendingRequests.get(msg.id);
-            if (pending) {
-              this.pendingRequests.delete(msg.id);
-              this.updateWorkerRef();
-              pending.resolve(msg.vectors);
-            }
-            break;
-          }
-          case "error": {
-            const pending = this.pendingRequests.get(msg.id);
-            if (pending) {
-              this.pendingRequests.delete(msg.id);
-              this.updateWorkerRef();
-              // If the worker reports a WASM-fatal or OOM error, reject with
-              // LocalProviderUnavailableError so callers (embed() → isAvailable)
-              // treat the local provider as broken and degrade to FTS-only.
-              // A generic Error would bypass that path, causing silent data loss.
-              // Uses the same isWasmFatalError() from embedding-worker-types.ts
-              // that the worker uses — single source of truth for classification.
-              if (isWasmFatalError(msg.error)) {
-                latchLocalProvider("terminal");
-                pending.reject(new LocalProviderUnavailableError(msg.error));
-              } else {
-                pending.reject(
-                  new Error(`Worker embedding failed: ${msg.error}`),
-                );
-              }
-            }
-            break;
-          }
-          case "init-needs-wasm": {
-            // The native ONNX backend loaded but couldn't parse an intact model
-            // (#1379 — Bun ↔ onnxruntime-node). Respawn a FRESH worker forcing
-            // WASM; the backend is committed per module-graph, so only a new
-            // worker can switch. One-shot: if WASM ALSO reports this (shouldn't
-            // happen — it means genuine corruption on both backends), fall
-            // through to the init-error path so we don't loop respawns.
-            if (!this.wasmFallbackTried) {
-              this.wasmFallbackTried = true;
-              this.forceWasm = true;
-              this.workerReady = false;
-              // Clear any transient init debt so the deliberate respawn isn't
-              // fast-failed, mirroring the OOM backoff path.
-              this.workerInitError = null;
-              log.info(
-                "native ONNX runtime could not load the embedding model " +
-                  `(${msg.error}); retrying with the bundled WASM runtime`,
-              );
-              void this.respawnForWasm();
-              break;
-            }
-            // Already tried WASM and it also failed — treat as a real init
-            // failure (genuine corruption on both backends, or a non-Bun cause).
-            this.handleInitError(msg.error);
-            break;
-          }
-          case "init-error": {
-            this.handleInitError(msg.error);
-            break;
-          }
-        }
-      });
-
-      // Worker crash / exit — reject all in-flight requests. Keep an errored
-      // worker owned until exit/shutdown so retirement can drain it completely.
-      let workerErrorHandled = false;
-      this.worker.prependOnceListener("error", () => {
-        workerErrorHandled = true;
-      });
-      this.worker.on("error", (err: Error) => {
-        if (this.worker !== spawned) return; // superseded worker — ignore
-        if (this.closing) return;
-        this.workerReady = false;
-        this.initPromise = null;
-        log.error("embedding worker crashed:", err);
-        // Worker errors can be transient (runtime/model startup pressure), so
-        // give them the same bounded cooldown/retry treatment as init-error.
-        // Keep the worker handle until exit/shutdown so pool retirement can
-        // drain it. The exit handler recognizes this already-classified error.
-        this.handleInitError(err.message);
-      });
-
-      this.worker.on("exit", (code) => {
-        if (this.worker !== spawned) return; // superseded worker — ignore
-        this.workerReady = false;
-        this.worker = null;
-        this.initPromise = null;
-
-        // Input-size-driven OOM: recover by respawning at a lower token cap on
-        // a fresh WASM heap (an in-process retry can't — WASM memory never
-        // shrinks), then re-submit the in-flight requests. Bounded: the cap
-        // backs off ×0.7 down to MIN_EMBED_TOKENS, at which point we latch
-        // FTS-only — so this can never loop unbounded (no event storm).
-        if (code === EMBED_OOM_EXIT_CODE) {
-          this.handleOomBackoff();
-          return;
-        }
-
-        // Any other non-zero exit is a genuine fatal crash — latch the
-        // provider broken so future ensureWorker() calls fast-fail instead of
-        // respawning a worker that will just crash again (event-storm guard).
-        if (code !== 0 && !workerErrorHandled && !this.closing) {
-          if (!this.workerInitError) {
-            this.workerInitError = `embedding worker exited with code ${code}`;
-            log.error(this.workerInitError, new Error(this.workerInitError));
-          }
-          latchLocalProvider("terminal");
-        }
-        for (const [, p] of this.pendingRequests) {
-          p.reject(
-            new LocalProviderUnavailableError(
-              this.workerInitError ?? "embedding worker exited",
-            ),
-          );
-        }
-        this.pendingRequests.clear();
-        if (!this.closing) this.onUnavailable?.();
-      });
-
-      this.workerReady = true;
-    })().catch((err) => {
-      this.initPromise = null; // allow retry
-      if (err instanceof LocalProviderUnavailableError) throw err;
-      const errorMsg = err instanceof Error ? err.message : String(err);
-      this.handleInitError(errorMsg);
-      throw new LocalProviderUnavailableError(errorMsg);
-    });
-
-    return this.initPromise;
-  }
-
-  /** Keep the worker ref'd while requests are in flight so the event loop
-   *  doesn't exit before responses arrive. When the pending map drains,
-   *  unref again so the worker doesn't prevent graceful process exit. */
-  private updateWorkerRef(): void {
-    if (!this.worker) return;
-    if (this.pendingRequests.size > 0) {
-      this.worker.ref();
-    } else {
-      this.worker.unref();
-    }
-  }
-
-  /** Throttled upward re-probe: when free memory has recovered past
-   *  EMBED_REPROBE_RATIO × the level at which the current cap was learned, step
-   *  the cap up one notch (bounded by the memory model for the now-larger pool)
-   *  and persist. Applied via the per-request cap, so it takes effect on the
-   *  next request without respawning the worker. An optimistic step is
-   *  corrected by the OOM backoff. */
-  private maybeReprobeCap(): void {
-    const now = Date.now();
-    if (now - this.lastReprobeAt < EMBED_REPROBE_INTERVAL_MS) return;
-    this.lastReprobeAt = now;
-    const free = containerFreeBytes();
-    if (!shouldReprobeEmbedCap(free, this.capFreememAtLearn)) return;
-    // The re-probe ceiling is the memory model over this worker's pool SHARE of
-    // free memory, so climbing back up never lets N workers sum past the host's
-    // RAM. The ratio check above stays on undivided free (it only compares
-    // now-vs-learn, both undivided).
-    const next = reprobeEmbedCap(
-      this.maxTokens,
-      free / this.memDivisor,
-      this.lastOomCap,
-    );
-    if (next <= this.maxTokens) return;
-    const prev = this.maxTokens;
-    this.maxTokens = next;
-    this.capFreememAtLearn = free;
-    persistEmbedCap(next, free, this.lastOomCap);
-    log.info(
-      `embedding cap re-probed up: ≤${prev} → ≤${next} tokens (free memory recovered)`,
-    );
-  }
-
-  /** Aggregate OOM context across in-flight requests for telemetry. We don't
-   *  know which request OOMed (the worker exits without posting), so report the
-   *  worst-case longest text and the in-flight count. */
-  private pendingOomContext(): { batchSize: number; longestChars: number } {
-    let batchSize = 0;
-    let longestChars = 0;
-    for (const [, p] of this.pendingRequests) {
-      batchSize++;
-      for (const t of p.payload.texts) {
-        if (t.length > longestChars) longestChars = t.length;
-      }
-    }
-    return { batchSize, longestChars };
-  }
-
-  /**
-   * Handle an OOM-signalled worker exit: lower the token cap ×0.7, persist it,
-   * and respawn (fresh heap) to re-run the in-flight requests. At the floor we
-   * give up and latch FTS-only. The worker is already nulled by the exit
-   * handler before this runs.
-   */
-  private handleOomBackoff(): void {
-    const capBefore = this.maxTokens;
-    const { batchSize, longestChars } = this.pendingOomContext();
-
-    if (capBefore <= MIN_EMBED_TOKENS) {
-      // Already at the floor and still OOMing → the host genuinely can't run
-      // local embeddings (system-wide exhaustion, not input size). Latch
-      // FTS-only and surface the remote-provider hint.
-      latchLocalProvider("terminal");
-      if (!localProviderErrorLogged) {
-        localProviderErrorLogged = true;
-        log.error(
-          `local embedding provider out of memory even at the ${MIN_EMBED_TOKENS}-token floor — ` +
-            `degrading to FTS-only search. Set search.embeddings.provider to 'voyage' or 'openai' ` +
-            `in .lore.json (with VOYAGE_API_KEY / OPENAI_API_KEY) for a remote provider.`,
-          new Error("embedding OOM at floor"),
-        );
-      }
-      fireEmbeddingFailure({
-        kind: "floor-latch",
-        capBefore,
-        capAfter: capBefore,
-        batchSize,
-        longestChars,
-      });
-      for (const [, p] of this.pendingRequests) {
-        p.reject(
-          new LocalProviderUnavailableError(
-            "embedding worker out of memory at the token floor",
-          ),
-        );
-      }
-      this.pendingRequests.clear();
-      return;
-    }
-
-    const free = containerFreeBytes();
-    const capAfter = backoffEmbedCap(capBefore);
-    this.maxTokens = capAfter;
-    // Remember the cap that just OOMed so the upward re-probe never climbs back
-    // to or past it within this process (a rising freemem doesn't prove the
-    // heap can grow that far).
-    this.lastOomCap = Math.max(this.lastOomCap, capBefore);
-    // Anchor the re-probe baseline at OOM-time free memory: only climb back up
-    // once memory has genuinely recovered (≥ EMBED_REPROBE_RATIO × this).
-    this.capFreememAtLearn = free;
-    // Persist the known-bad cap alongside the backed-off cap so the NEXT process
-    // start won't re-probe up to it even if the box reboots with more free RAM.
-    persistEmbedCap(capAfter, free, this.lastOomCap);
-    log.info(
-      `embedding worker OOM at ≤${capBefore} tokens — backing off to ≤${capAfter} ` +
-        `and respawning on a fresh heap (${batchSize} in-flight, longest≈${longestChars} chars)`,
-    );
-    fireEmbeddingFailure({
-      kind: "oom-backoff",
-      capBefore,
-      capAfter,
-      batchSize,
-      longestChars,
-    });
-
-    // Clear any stale worker error so the deliberate respawn isn't fast-failed
-    // by ensureWorker() (mirrors shutdown()). The OOM exit is recoverable — a
-    // prior `error`/`init-error` could otherwise leave workerInitError set and
-    // block recovery.
-    this.workerInitError = null;
-
-    // Respawn and re-submit. Fire-and-forget: ensureWorker()'s own handlers
-    // reject pending if the respawn itself fails.
-    void this.resubmitPending();
-  }
-
-  /** Respawn the worker (at the already-lowered cap) and re-post every
-   *  in-flight request so the OOM backoff is transparent to callers. */
-  private async resubmitPending(): Promise<void> {
-    if (this.pendingRequests.size === 0) return;
-    try {
-      await this.ensureWorker();
-    } catch {
-      // A *synchronous* respawn failure (e.g. `new Worker(...)` throws) rejects
-      // initPromise before any worker event handler is attached, so nothing
-      // else will ever settle these requests. Reject them here to avoid a hung
-      // caller — the local embed path has no timeout. (Asynchronous spawn
-      // failures are already settled by the worker error/exit/init-error
-      // handlers, which clear the map, so this loop is then a no-op.)
-      for (const [, p] of this.pendingRequests) {
-        p.reject(
-          new LocalProviderUnavailableError(
-            "embedding worker respawn failed after OOM backoff",
-          ),
-        );
-      }
-      this.pendingRequests.clear();
-      return;
-    }
-    const worker = this.worker;
-    if (!worker) return; // raced with another exit — that handler owns pending
-    for (const [, p] of this.pendingRequests) {
-      // Re-submit at the lowered cap so the retry doesn't re-OOM at the old one,
-      // further clamped to live free memory for this worker's pool share.
-      p.payload.maxTokens = this.effectiveMaxTokens();
-      try {
-        worker.postMessage(p.payload satisfies WorkerInbound);
-      } catch (error) {
-        const errorMsg =
-          error instanceof Error
-            ? error.message
-            : "embedding worker died during OOM resubmit";
-        this.handleInitError(errorMsg);
-        return;
-      }
-    }
-    this.updateWorkerRef();
-  }
-
-  /**
-   * Shared handler for a fatal worker init failure. Surfaces
-   * `LocalProviderUnavailableError` on all pending + future requests, either
-   * latching FTS-only (missing stack / retry budget exhausted) or scheduling a
-   * cooldown retry with a fresh worker. Called from the `init-error` message and
-   * from `init-needs-wasm` once the one-shot WASM fallback has already been used
-   * (a WASM init that also fails is a genuine failure, #1379).
-   */
-  private handleInitError(errorMsg: string): void {
-    this.workerInitError = errorMsg;
-    this.workerReady = false;
-    if (isMissingLocalStackError(errorMsg)) {
-      // Optional local-embedding stack not installed (#1026 — an expected,
-      // actionable degraded state that will NOT self-heal on its own). Latch
-      // permanently; recall degrades to FTS-only. Flag it so the self-heal
-      // re-probe skips it (a reinstall, not a retry, is what recovers this).
-      latchLocalProvider("terminal");
-      localStackMissing = true;
-      if (!localProviderErrorLogged) {
-        localProviderErrorLogged = true;
-        log.warn(
-          "local embedding dependencies not installed " +
-            "(optional '@huggingface/transformers' / 'onnxruntime-node' absent) — " +
-            "recall will use FTS-only search. Reinstall without --omit=optional to " +
-            "enable local embeddings, or set search.embeddings.provider in .lore.json " +
-            "to a remote provider (voyage/openai) with the matching API key.",
-        );
-      }
-    } else {
-      // A potentially transient failure (e.g. a model read that raced a
-      // concurrent writer during a multi-instance restart, momentary memory
-      // pressure). Retry a FRESH worker after a cooldown rather than disabling
-      // local embeddings for the whole process lifetime; only give up (latch)
-      // once the retry budget is exhausted.
-      localInitFailures++;
-      localInitFailureGeneration++;
-      if (localInitFailures >= LOCAL_INIT_MAX_ATTEMPTS) {
-        latchLocalProvider("transient-init-exhausted");
-        localInitRetryAt = 0;
-        if (!localProviderErrorLogged) {
-          localProviderErrorLogged = true;
-          log.error(
-            `local embedding provider failed to init after ${localInitFailures} attempts: ${errorMsg}. ` +
-              `Set search.embeddings.provider in .lore.json to use a remote provider.`,
-            new Error(`embedding worker init failed: ${errorMsg}`),
-          );
-        }
-      } else {
-        const retryDelayMs = computeInitRetryDelayMs(
-          localInitFailures,
-          localInitCooldownMs,
-        );
-        localInitRetryAt = Date.now() + retryDelayMs;
-        log.warn(
-          `local embedding init failed (attempt ${localInitFailures}/${LOCAL_INIT_MAX_ATTEMPTS}): ${errorMsg}. ` +
-            `Retrying with a fresh worker in ~${Math.round(retryDelayMs / 1000)}s; recall is FTS-only until then.`,
-        );
-      }
-    }
-    for (const [, p] of this.pendingRequests) {
-      p.reject(new LocalProviderUnavailableError(errorMsg));
-    }
-    this.pendingRequests.clear();
-    this.updateWorkerRef();
-    this.onUnavailable?.();
-  }
-
-  /**
-   * Respawn a FRESH worker forcing the bundled WASM ONNX Runtime after a native
-   * worker reported `init-needs-wasm` (#1379). The native worker is still alive
-   * (it returned from init without exiting), so terminate it first, then let
-   * `ensureWorker()` build a new one — `this.forceWasm` (set by the caller) makes
-   * its `workerData` skip native. Any request that triggered the failed init is
-   * re-posted to the WASM worker so the fallback is transparent to callers.
-   */
-  private async respawnForWasm(): Promise<void> {
-    const dead = this.worker;
-    this.worker = null;
-    this.initPromise = null;
-    // Belt-and-suspenders: ensureWorker() early-returns when `workerReady` is
-    // true, so it MUST be false here or the WASM worker would never spawn and
-    // pending requests would be lost. The `init-needs-wasm` caller already
-    // clears it, but re-clear locally so this invariant holds for any caller and
-    // survives refactors (defense against the workerReady race Seer flagged).
-    this.workerReady = false;
-    if (dead) {
-      // The fresh worker is independent, so fallback need not wait for a slow
-      // native teardown. Keep owning the old generation, though: final process
-      // shutdown must join it and must fail closed if Node cannot confirm exit.
-      // NOTE: terminate() DOES emit an async `exit(1)` on the dead worker. That
-      // event is harmless here only because each handler is bound to its own
-      // `spawned` worker and early-returns when `this.worker` has moved on (see
-      // ensureWorker) — otherwise the stale exit would latch the provider broken
-      // and clobber the fresh WASM worker (#1387-B1).
-      this.trackRetiredWorkerTermination(dead);
-    }
-    try {
-      await this.ensureWorker();
-    } catch {
-      // Synchronous respawn failure — nothing else will settle these requests
-      // (the local embed path has no timeout). Reject them here.
-      for (const [, p] of this.pendingRequests) {
-        p.reject(
-          new LocalProviderUnavailableError(
-            "embedding worker respawn failed after WASM fallback",
-          ),
-        );
-      }
-      this.pendingRequests.clear();
-      return;
-    }
-    // Re-read through a cast: TS still has `this.worker` narrowed to null from
-    // the `this.worker = null` above (it can't see that ensureWorker() reassigns
-    // it across the await), so a direct read would be typed `never`.
-    const worker = this.worker as import("node:worker_threads").Worker | null;
-    if (worker == null) return; // raced with an exit — that handler owns pending
-    for (const [, p] of this.pendingRequests) {
-      p.payload.maxTokens = this.effectiveMaxTokens();
-      try {
-        worker.postMessage(p.payload satisfies WorkerInbound);
-      } catch (error) {
-        const errorMsg =
-          error instanceof Error
-            ? error.message
-            : "embedding worker died during WASM resubmit";
-        this.handleInitError(errorMsg);
-        return;
-      }
-    }
-    this.updateWorkerRef();
-  }
-
-  private trackRetiredWorkerTermination(worker: ShutdownableWorker): void {
-    let termination: Promise<number>;
-    try {
-      termination = worker.terminate();
-    } catch (error) {
-      termination = Promise.reject(error);
-    }
-    let tracked: Promise<void>;
-    tracked = termination
-      .then(
-        () => {},
-        (error: unknown) => {
-          this.retiredWorkerShutdownErrors.push(error);
-        },
-      )
-      .finally(() => this.retiredWorkerShutdowns.delete(tracked));
-    this.retiredWorkerShutdowns.add(tracked);
-  }
-
-  private async settleRetiredWorkers(timeoutMs: number): Promise<void> {
-    const deadlineAt = Date.now() + Math.max(0, timeoutMs);
-    while (this.retiredWorkerShutdowns.size > 0) {
-      const remainingMs = Math.max(0, deadlineAt - Date.now());
-      let timer: ReturnType<typeof setTimeout> | undefined;
-      try {
-        await Promise.race([
-          Promise.all(this.retiredWorkerShutdowns),
-          new Promise<never>((_resolve, reject) => {
-            timer = setTimeout(
-              () =>
-                reject(
-                  new Error(
-                    "retired embedding worker did not settle before shutdown deadline",
-                  ),
-                ),
-              remainingMs,
-            );
-            timer.unref?.();
-          }),
-        ]);
-      } finally {
-        if (timer) clearTimeout(timer);
-      }
-    }
-    if (this.retiredWorkerShutdownErrors.length > 0) {
-      throw new AggregateError(
-        this.retiredWorkerShutdownErrors,
-        "embedding worker termination was not confirmed",
-      );
-    }
-  }
-
-  async embed(
-    texts: string[],
-    inputType: "document" | "query",
-    signal?: AbortSignal,
-    onExecutionStart?: () => void,
-  ): Promise<Float32Array[]> {
-    if (signal?.aborted) throw new EmbeddingRequestAbortedError();
-    await this.ensureWorker();
-    const worker = this.worker;
-    if (this.closing || !worker) {
-      throw new LocalProviderUnavailableError(
-        "embedding worker closed during initialization",
-      );
-    }
-    // Opportunistically raise the cap if free memory has recovered (cheap,
-    // throttled). Takes effect via the per-request cap below — no respawn.
-    this.maybeReprobeCap();
-
-    // Pre-truncate texts that exceed the safe ONNX inference limit.
-    // This prevents OOM on single inputs near the model's 8192-token max.
-    const truncated = texts.map(safeLocalTruncate);
-
-    // Prepend Nomic task instruction prefix.
-    const prefix =
-      inputType === "document" ? "search_document: " : "search_query: ";
-    const prefixed = truncated.map((t) => prefix + t);
-
-    const id = this.nextRequestId++;
-    // Recall queries (single query-type texts) get high priority so they
-    // jump ahead of any queued backfill batches in the worker.
-    const priority = isRecallEmbed(texts, inputType) ? "high" : "normal";
-
-    const payload: EmbedRequest = {
-      type: "embed",
-      id,
-      texts: prefixed,
-      inputType,
-      priority,
-      // Clamp to what CURRENT free memory allows for this worker's pool share,
-      // not just the construction-time learned cap — free memory drops as
-      // sibling workers and live sessions allocate, and a native over-allocation
-      // is an uncatchable SIGKILL rather than a recoverable in-worker OOM.
-      maxTokens: this.effectiveMaxTokens(),
-    };
-
-    return new Promise<Float32Array[]>((resolve, reject) => {
-      let settled = false;
-      const cleanup = (): void => {
-        if (signal) signal.removeEventListener("abort", onAbort);
-      };
-      const resolveRequest = (vectors: Float32Array[]): void => {
-        if (settled) return;
-        settled = true;
-        cleanup();
-        resolve(vectors);
-      };
-      const rejectRequest = (error: Error): void => {
-        if (settled) return;
-        settled = true;
-        cleanup();
-        reject(error);
-      };
-      const onAbort = (): void => {
-        if (!this.pendingRequests.delete(id)) return;
-        this.updateWorkerRef();
-        rejectRequest(new EmbeddingRequestAbortedError());
-      };
-      this.pendingRequests.set(id, {
-        resolve: resolveRequest,
-        reject: rejectRequest,
-        payload,
-        onExecutionStart,
-      });
-      signal?.addEventListener("abort", onAbort, { once: true });
-      if (signal?.aborted) {
-        onAbort();
-        return;
-      }
-      this.updateWorkerRef();
-      try {
-        worker.postMessage(payload satisfies WorkerInbound);
-      } catch {
-        // Worker may have been terminated between ensureWorker() and here
-        // (race with process.exit(1) in the worker thread). Clean up and
-        // reject with the expected error type so callers degrade gracefully.
-        this.handleInitError(
-          "embedding worker terminated before request could be sent",
-        );
-      }
-    });
-  }
-
-  /** Shut down the worker thread. Called by `resetProvider()` on config change.
-   *  Sends a shutdown message so the worker drains in-flight work and exits
-   *  via a deferred `process.exit(0)` (lets NAPI callbacks unwind safely).
-   *
-   *  Returns a promise that resolves once the worker has fully exited. Callers
-   *  that need a clean teardown (tests, config change) should await the result.
-   *  Fire-and-forget callers (process exit) can ignore it. */
-  shutdown(timeoutMs = WORKER_SHUTDOWN_TIMEOUT_MS): Promise<void> {
-    if (this.shutdownPromise) return this.shutdownPromise;
-    this.closing = true;
-    this.workerReady = false;
-    this.workerInitError = null;
-
-    // Reject any in-flight requests with LocalProviderUnavailableError so
-    // fire-and-forget callers' catch blocks handle it the same way as other
-    // provider failures (graceful degradation, no Sentry noise).
-    for (const [, p] of this.pendingRequests) {
-      p.reject(new LocalProviderUnavailableError("embedding worker shut down"));
-    }
-    this.pendingRequests.clear();
-
-    const init = this.initPromise;
-    this.shutdownPromise = (async () => {
-      await init?.catch(() => {});
-      const worker = this.worker;
-      this.worker = null;
-      this.initPromise = null;
-      const outcomes = await Promise.allSettled([
-        (async () => {
-          if (!worker) return;
-          // Don't let a mid-backfill ref keep the event loop alive while we wait
-          // for the worker to exit.
-          worker.unref();
-          await awaitWorkerShutdown(worker, timeoutMs);
-        })(),
-        this.settleRetiredWorkers(timeoutMs),
-      ]);
-      const failures = outcomes.flatMap((outcome) =>
-        outcome.status === "rejected" ? [outcome.reason] : [],
-      );
-      if (failures.length > 0) {
-        if (failures.length === 1) throw failures[0];
-        throw new AggregateError(
-          failures,
-          "embedding worker termination was not confirmed",
-        );
-      }
-    })();
-    return this.shutdownPromise;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Local embedding worker pool (#999)
-// ---------------------------------------------------------------------------
-
-/** Resolve the configured embedding-pool ceiling: `LORE_EMBED_POOL_SIZE` wins
- *  (the escape hatch — `=1` forces today's single worker), then
- *  `search.embeddings.embedPoolSize`, else `undefined` (memory-gated default).
- *  Read per-construction (not cached) so config reloads / env changes take
- *  effect on the next provider. Invalid values are ignored (fall through). */
-function configuredEmbedPoolSize(): number | undefined {
-  const raw = process.env.LORE_EMBED_POOL_SIZE;
-  if (raw !== undefined) {
-    const n = Number(raw);
-    if (Number.isFinite(n) && n >= 1) return Math.floor(n);
-    // invalid env → ignore, fall through to config
-  }
-  const cfg = config().search.embeddings.embedPoolSize;
-  if (typeof cfg === "number" && Number.isFinite(cfg) && cfg >= 1) {
-    return Math.floor(cfg);
-  }
-  return undefined;
-}
-
-/** Test seam: exposes {@link configuredEmbedPoolSize} so suites can assert the
- *  env/config resolution + invalid-value fall-through (invalid env must resolve
- *  to `undefined`, never `NaN`) without spinning up a pool. */
-export function _configuredEmbedPoolSize(): number | undefined {
-  return configuredEmbedPoolSize();
-}
-
-/** Resolve the legacy backfill CPU-duty setting for configuration compatibility.
- *  The durable temporal scheduler now bounds inference directly, so this value
- *  no longer throttles temporal backfill work. Invalid values fall through. */
-function configuredBackfillCpuDuty(): number | undefined {
-  const raw = process.env.LORE_BACKFILL_CPU_DUTY;
-  if (raw !== undefined) {
-    const n = Number(raw);
-    if (Number.isFinite(n) && n > 0) return n; // clamped in resolveBackfillCpuDuty
-    // invalid env → ignore, fall through to config
-  }
-  const cfg = config().search.embeddings.backfillCpuDuty;
-  if (typeof cfg === "number" && Number.isFinite(cfg) && cfg > 0) return cfg;
-  return undefined;
-}
-
-/** Test seam for the retained legacy configuration resolution. */
-export function _configuredBackfillCpuDuty(): number | undefined {
-  return configuredBackfillCpuDuty();
-}
-
-/** Test-only override of the embedding-pool ceiling (null clears). Sets the
- *  construction-time ceiling directly, bypassing the freemem-based sizing so
- *  suites don't depend on the host's actual RAM. */
-let testEmbedPoolSize: number | null = null;
-export function _setEmbedPoolSizeForTest(n: number | null): void {
-  testEmbedPoolSize = n;
-}
-
-/** Test-only override for the pool's LIVE per-spawn memory gate (null → real
- *  `os.freemem()`). Lets suites drive the "grow when memory allows" vs "stay at
- *  one worker when memory is tight" paths deterministically. */
-let testPoolFreememBytes: number | null = null;
-export function _setPoolFreememForTest(bytes: number | null): void {
-  testPoolFreememBytes = bytes;
-}
-
-/** One worker slot and its actual execution count. Host queueing keeps this at
- * zero or one. It stays occupied after caller cancellation and across an OOM
- * respawn until the underlying provider operation settles. */
-interface EmbedSlot {
-  provider: LocalProvider;
-  inflight: number;
-  healthy: boolean;
-  recoveryGeneration: number | null;
-  retirement: Promise<void> | null;
-}
-
-type PoolOperationState = "queued" | "running" | "completed";
-
-interface PoolWaiter {
-  settled: boolean;
-  signal?: AbortSignal;
-  onAbort?: () => void;
-  resolve: (vectors: Float32Array[]) => void;
-  reject: (error: Error) => void;
-}
-
-interface PoolOperation {
-  key: string;
-  texts: string[];
-  byteSize: number;
-  inputType: "document" | "query";
-  priority: "high" | "normal";
-  state: PoolOperationState;
-  waiters: Set<PoolWaiter>;
-  retainResult?: boolean;
-  vectors?: Float32Array[];
-  completedAt?: number;
-}
-
-/** Successfully completed prefix of one interrupted token-batched call. The
- * next durable temporal retry consumes it once and resumes at `nextIndex`. */
-interface TokenBatchCheckpoint {
-  nextIndex: number;
-  vectors: Float32Array[];
-  updatedAt: number;
-}
-
-/** Keep a successful orphan result long enough for a durable scheduler's
- * bounded retry to reclaim it without running the same native inference twice. */
-const COMPLETED_EMBED_REUSE_MS = 5 * 60_000;
-const MAX_COMPLETED_EMBED_RESULTS = 8;
-/** Bound raw input retained by the host queue and the O(n) cancellation scan.
- * Normal background work stops below the hard ceiling so latency-sensitive
- * recall can still enter a saturated queue. In-flight work is separately
- * bounded by the four-worker absolute pool ceiling. */
-const MAX_QUEUED_EMBED_OPERATIONS = 256;
-const MAX_QUEUED_EMBED_BYTES = 8 * 1024 * 1024;
-const MAX_NORMAL_QUEUED_EMBED_OPERATIONS = 240;
-const MAX_NORMAL_QUEUED_EMBED_BYTES = 7 * 1024 * 1024;
-/** Deduplication must not turn one bounded operation into an unbounded listener
- * set under a same-input flood. The operation cap bounds the aggregate. */
-const MAX_WAITERS_PER_EMBED_OPERATION = 64;
-const DEFAULT_EMBED_INIT_WATCHDOG_MS = 10 * 60_000;
-const DEFAULT_EMBED_EXECUTION_WATCHDOG_MS = 5 * 60_000;
-let embedInitWatchdogMs = DEFAULT_EMBED_INIT_WATCHDOG_MS;
-let embedExecutionWatchdogMs = DEFAULT_EMBED_EXECUTION_WATCHDOG_MS;
-
-/** Test seam for independent worker-owned watchdogs. */
-export function _setEmbeddingWorkerWatchdogsForTest(
-  initMs: number | null,
-  executionMs: number | null,
-): void {
-  embedInitWatchdogMs = initMs ?? DEFAULT_EMBED_INIT_WATCHDOG_MS;
-  embedExecutionWatchdogMs = executionMs ?? DEFAULT_EMBED_EXECUTION_WATCHDOG_MS;
-}
-
-function embeddingOperationKey(
-  texts: string[],
-  inputType: "document" | "query",
-): string {
-  const hash = createHash("sha256");
-  hash.update(inputType);
-  for (const value of texts) {
-    hash.update("\0");
-    hash.update(String(Buffer.byteLength(value)));
-    hash.update(":");
-    hash.update(value);
-  }
-  return hash.digest("hex");
-}
-
-function cloneEmbeddingVectors(vectors: Float32Array[]): Float32Array[] {
-  return vectors.map((vector) => vector.slice());
-}
-
-/** Internal dispatch signal: the provider can recover at `retryAt`, so queued
- * work must remain owned instead of being rejected as permanently unavailable. */
-class EmbeddingWorkerRetryCooldownError extends LocalProviderUnavailableError {
-  readonly retryAt: number;
-
-  constructor(retryAt: number) {
-    super("embedding worker retry cooldown is active");
-    this.name = "EmbeddingWorkerRetryCooldownError";
-    this.retryAt = retryAt;
-  }
-}
-
-/**
- * A pool of {@link LocalProvider} workers so concurrent embeds run in parallel
- * instead of serializing through a single worker (#999). Mirrors
- * `vector-pool.ts`, but over the stateful embedding provider rather than raw
- * workers — each LocalProvider keeps its full, tested per-worker lifecycle (OOM
- * ×0.7 backoff + respawn + resubmit, corrupt-model self-heal, WASM-fatal latch,
- * cap re-probe, bounded shutdown).
- *
- * The host owns a priority queue and posts at most one request to each worker.
- * A cancelled queued request is removed without reaching the worker. A caller
- * that cancels running work detaches promptly while execution remains accounted
- * for until completion or a worker-owned watchdog retires the slot. Identical
- * retries join running work or reclaim its bounded cached result.
- *
- * Growth is LAZY and MEMORY-GATED: worker 0 spawns on the first embed (today's
- * behavior); a secondary spawns only under genuine concurrent demand, below the
- * ceiling, AND when a full {@link PER_WORKER_MEM_BUDGET_BYTES} is free at spawn
- * time — so a light or constrained host never loads a second ~680 MB model. The
- * Terminal failures (missing stack / WASM-fatal / floor-OOM) latch the whole
- * provider. Transient init failures retire only their slot; proven siblings stay
- * usable while replacement spawning observes the bounded retry cooldown.
- */
-class EmbeddingPool implements EmbeddingProvider {
-  readonly maxBatchSize = 256;
-
-  private readonly modelId: string;
-  private readonly dimensions: number;
-  /** Memory-gated ceiling, fixed at construction (durable — no mid-session
-   *  recompute). Live freemem re-gates each actual spawn on top of this. */
-  private readonly ceiling: number;
-  private readonly slots: EmbedSlot[] = [];
-  private readonly retiredShutdowns = new Set<Promise<void>>();
-  private readonly retiredShutdownErrors: unknown[] = [];
-  /** Queue on the host so at most one native batch is posted to each worker.
-   * Caller cancellation can then detach without changing worker lifetime. */
-  private readonly queue: PoolOperation[] = [];
-  private queuedBytes = 0;
-  private readonly operations = new Map<string, PoolOperation>();
-  private readonly tokenBatchCheckpoints = new Map<
-    string,
-    TokenBatchCheckpoint
-  >();
-  private dispatching = false;
-  private retryDispatchTimer: ReturnType<typeof setTimeout> | null = null;
-  private retryDispatchAt = 0;
-  private closing = false;
-  private shutdownPromise: Promise<void> | null = null;
-
-  constructor(modelId: string, dimensions: number) {
-    this.modelId = modelId;
-    this.dimensions = dimensions;
-    if (testEmbedPoolSize != null) {
-      // Deterministic test override — bypass the memory gate entirely.
-      this.ceiling = Math.max(
-        1,
-        Math.min(Math.floor(testEmbedPoolSize), EMBED_POOL_ABS_MAX),
-      );
-    } else if (process.env.NODE_ENV === "test") {
-      // Keep existing single-worker suites deterministic regardless of CI RAM:
-      // honor an explicit config/env ceiling (clamped like the prod branch),
-      // else default to one worker.
-      this.ceiling = Math.max(
-        1,
-        Math.min(configuredEmbedPoolSize() ?? 1, EMBED_POOL_ABS_MAX),
-      );
-    } else {
-      this.ceiling = desiredEmbedPoolSize(
-        this.liveFreemem(),
-        configuredEmbedPoolSize(),
-      );
-    }
-  }
-
-  /** Pick the slot to dispatch the next request to, growing the pool lazily
-   *  when there's concurrent demand, headroom below the ceiling, and memory. */
-  private pickSlot(): EmbedSlot {
-    const now = Date.now();
-    const healthySlots = this.slots.filter((slot) => slot.healthy);
-    const terminallyUnavailable =
-      localProviderFailureCause === "terminal" ||
-      (localProviderFailureCause === "transient-init-exhausted" &&
-        healthySlots.length === 0);
-    if (this.closing || terminallyUnavailable) {
-      throw new LocalProviderUnavailableError("embedding pool is unavailable");
-    }
-    // Primary worker: always present (its own OOM backoff, not pool sizing,
-    // protects a constrained host — identical to today's single worker).
-    if (this.slots.length === 0) {
-      if (localInitRetryAt > now) {
-        throw new EmbeddingWorkerRetryCooldownError(localInitRetryAt);
-      }
-      // Admit exactly one recovery probe after the cooldown. The outstanding
-      // failure debt below prevents pool growth until this slot succeeds.
-      if (localInitRetryAt > 0) localInitRetryAt = 0;
-      return this.spawnSlot(localInitFailures > 0);
-    }
-
-    // During transient failure debt, route new work only to proven siblings.
-    // Unproven slots may still finish their existing requests, but cannot
-    // amplify a bad generation with more work or replacement spawns.
-    const eligible =
-      healthySlots.length > 0 &&
-      (localInitFailures > 0 ||
-        localInitRetryAt > 0 ||
-        localProviderFailureCause === "transient-init-exhausted")
-        ? healthySlots
-        : this.slots;
-    let best = eligible[0];
-    for (const s of eligible) {
-      if (s.inflight < best.inflight) best = s;
-    }
-
-    // Every worker is busy: add capacity only after one slot has completed a
-    // real embed. Cold workers share the same HuggingFace cache; starting two
-    // before either is healthy lets one read/purge the other's partial download.
-    // Once bootstrap succeeds, retain the normal lazy, memory-gated growth.
-    const canGrow =
-      healthySlots.length > 0 &&
-      best.inflight > 0 &&
-      this.slots.length + this.retiredShutdowns.size < this.ceiling &&
-      this.liveFreemem() >= PER_WORKER_MEM_BUDGET_BYTES;
-    if (canGrow) {
-      if (localInitRetryAt > 0) {
-        if (localInitRetryAt > now) return best;
-        // The cooldown admits one recovery slot. Failure re-arms the next
-        // backoff; success clears all transient debt.
-        localInitRetryAt = 0;
-        return this.spawnSlot(true);
-      }
-      if (localInitFailures === 0) return this.spawnSlot(false);
-    }
-    return best;
-  }
-
-  /** Container-aware free memory for the live per-spawn gate: the injected test
-   *  value (or live `freemem()`) clamped to the cgroup limit, so a memory-capped
-   *  container never spawns a second native-ONNX worker off the host's (much
-   *  larger) free figure. Overridable in tests via {@link _setPoolFreememForTest}
-   *  (host free) and {@link _setConstrainedMemoryForTest} (the cgroup limit). */
-  private liveFreemem(): number {
-    const raw = testPoolFreememBytes != null ? testPoolFreememBytes : freemem();
-    return clampFreeToContainerLimit(raw, constrainedMemoryLimit());
-  }
-
-  private spawnSlot(recoveryProbe = false): EmbedSlot {
-    let slot: EmbedSlot;
-    const provider = new LocalProvider(
-      this.modelId,
-      this.dimensions,
-      this.ceiling,
-      () => {
-        void this.retireSlot(slot);
-      },
-    );
-    slot = {
-      // Pass the pool ceiling as the memory divisor so every worker sizes its
-      // token cap from `free / ceiling`. The ceiling is itself memory-gated
-      // (desiredEmbedPoolSize), so the workers the host is provisioned for
-      // collectively stay within one `EMBED_MEM_FRACTION` share of free memory
-      // instead of each independently claiming half and summing to an OOM.
-      provider,
-      inflight: 0,
-      healthy: false,
-      recoveryGeneration: recoveryProbe ? localInitFailureGeneration : null,
-      retirement: null,
-    };
-    this.slots.push(slot);
-    return slot;
-  }
-
-  private retireSlot(slot: EmbedSlot): Promise<void> {
-    if (slot.retirement) return slot.retirement;
-    const index = this.slots.indexOf(slot);
-    if (index === -1) return Promise.resolve();
-    this.slots.splice(index, 1);
-    this.preserveHealthyServiceAfterExhaustion();
-    const retirement = slot.provider.shutdown().catch((error: unknown) => {
-      // Runtime recovery must not create an unhandled rejection, but a later
-      // process/config shutdown still needs to know that this worker's exit was
-      // never confirmed.
-      this.retiredShutdownErrors.push(error);
-    });
-    slot.retirement = retirement;
-    this.retiredShutdowns.add(retirement);
-    void retirement.finally(() => {
-      this.retiredShutdowns.delete(retirement);
-      this.dispatch();
-    });
-    return retirement;
-  }
-
-  private preserveHealthyServiceAfterExhaustion(): void {
-    if (
-      localProviderFailureCause !== "transient-init-exhausted" ||
-      !this.hasHealthySlot()
-    ) {
-      return;
-    }
-    // The retry budget still stays exhausted, preventing another replacement
-    // storm. Only the global FTS-only latch is invalid when a sibling works.
-    clearLocalProviderLatch();
-    localProviderErrorLogged = false;
-  }
-
-  private settleWaiter(
-    operation: PoolOperation,
-    waiter: PoolWaiter,
-    outcome: { vectors: Float32Array[] } | { error: Error },
-  ): void {
-    if (waiter.settled) return;
-    waiter.settled = true;
-    operation.waiters.delete(waiter);
-    if (waiter.signal && waiter.onAbort) {
-      waiter.signal.removeEventListener("abort", waiter.onAbort);
-    }
-    if ("vectors" in outcome) {
-      waiter.resolve(cloneEmbeddingVectors(outcome.vectors));
-    } else {
-      waiter.reject(outcome.error);
-    }
-  }
-
-  private dropUnobservedQueuedOperation(operation: PoolOperation): void {
-    if (operation.state !== "queued" || operation.waiters.size > 0) return;
-    const index = this.queue.indexOf(operation);
-    if (index !== -1) this.removeQueuedOperation(index);
-    if (this.operations.get(operation.key) === operation) {
-      this.operations.delete(operation.key);
-    }
-    operation.texts = [];
-  }
-
-  private attachWaiter(
-    operation: PoolOperation,
-    signal?: AbortSignal,
-  ): Promise<Float32Array[]> {
-    if (signal?.aborted) {
-      return Promise.reject(new EmbeddingRequestAbortedError());
-    }
-    if (operation.state === "completed" && operation.vectors) {
-      const vectors = cloneEmbeddingVectors(operation.vectors);
-      if (this.operations.get(operation.key) === operation) {
-        this.operations.delete(operation.key);
-      }
-      operation.vectors = undefined;
-      return Promise.resolve(vectors);
-    }
-    if (operation.waiters.size >= MAX_WAITERS_PER_EMBED_OPERATION) {
-      return Promise.reject(new EmbeddingQueueCapacityError());
-    }
-
-    return new Promise<Float32Array[]>((resolve, reject) => {
-      const waiter: PoolWaiter = { settled: false, signal, resolve, reject };
-      const onAbort = (): void => {
-        this.settleWaiter(operation, waiter, {
-          error: new EmbeddingRequestAbortedError(),
-        });
-        if (operation.state === "running" && operation.waiters.size === 0) {
-          operation.retainResult = true;
-        }
-        this.dropUnobservedQueuedOperation(operation);
-      };
-      waiter.onAbort = onAbort;
-      operation.waiters.add(waiter);
-      signal?.addEventListener("abort", onAbort, { once: true });
-      if (signal?.aborted) onAbort();
-    });
-  }
-
-  private enqueueOperation(operation: PoolOperation): void {
-    this.queuedBytes += operation.byteSize;
-    if (operation.priority === "high") {
-      let insertAt = 0;
-      while (
-        insertAt < this.queue.length &&
-        this.queue[insertAt].priority === "high"
-      ) {
-        insertAt++;
-      }
-      this.queue.splice(insertAt, 0, operation);
-      return;
-    }
-    this.queue.push(operation);
-  }
-
-  private removeQueuedOperation(index: number): PoolOperation | undefined {
-    const [operation] = this.queue.splice(index, 1);
-    if (operation) {
-      this.queuedBytes = Math.max(0, this.queuedBytes - operation.byteSize);
-    }
-    if (this.queue.length === 0) this.clearRetryDispatchTimer();
-    return operation;
-  }
-
-  private clearRetryDispatchTimer(): void {
-    if (this.retryDispatchTimer) clearTimeout(this.retryDispatchTimer);
-    this.retryDispatchTimer = null;
-    this.retryDispatchAt = 0;
-  }
-
-  /** Resume the host queue when a transient provider cooldown expires. Keeping
-   * this timer pool-owned avoids a cascade of false terminal rejections while
-   * still bounding retained work through the normal queue limits. */
-  private scheduleRetryDispatch(retryAt: number): void {
-    if (this.retryDispatchTimer && this.retryDispatchAt === retryAt) return;
-    this.clearRetryDispatchTimer();
-    this.retryDispatchAt = retryAt;
-    const timer = setTimeout(
-      () => {
-        if (this.retryDispatchTimer !== timer) return;
-        this.retryDispatchTimer = null;
-        this.retryDispatchAt = 0;
-        this.dispatch();
-      },
-      Math.max(0, retryAt - Date.now()),
-    );
-    this.retryDispatchTimer = timer;
-  }
-
-  private canEnqueueOperation(
-    priority: PoolOperation["priority"],
-    byteSize: number,
-  ): boolean {
-    const maxOperations =
-      priority === "high"
-        ? MAX_QUEUED_EMBED_OPERATIONS
-        : MAX_NORMAL_QUEUED_EMBED_OPERATIONS;
-    const maxBytes =
-      priority === "high"
-        ? MAX_QUEUED_EMBED_BYTES
-        : MAX_NORMAL_QUEUED_EMBED_BYTES;
-    return (
-      byteSize <= maxBytes &&
-      this.queue.length < maxOperations &&
-      this.queuedBytes <= maxBytes - byteSize
-    );
-  }
-
-  private pruneCompletedOperations(now = Date.now()): void {
-    const completed: PoolOperation[] = [];
-    for (const [key, operation] of this.operations) {
-      if (operation.state !== "completed") continue;
-      if (
-        operation.completedAt === undefined ||
-        now - operation.completedAt > COMPLETED_EMBED_REUSE_MS
-      ) {
-        this.operations.delete(key);
-        operation.vectors = undefined;
-      } else {
-        completed.push(operation);
-      }
-    }
-    completed.sort((a, b) => (a.completedAt ?? 0) - (b.completedAt ?? 0));
-    while (completed.length > MAX_COMPLETED_EMBED_RESULTS) {
-      const operation = completed.shift();
-      if (!operation) break;
-      if (this.operations.get(operation.key) === operation) {
-        this.operations.delete(operation.key);
-      }
-      operation.vectors = undefined;
-    }
-  }
-
-  private pruneTokenBatchCheckpoints(now = Date.now()): void {
-    const retained = [...this.tokenBatchCheckpoints.entries()]
-      .filter(([, checkpoint]) => {
-        if (now - checkpoint.updatedAt <= COMPLETED_EMBED_REUSE_MS) return true;
-        checkpoint.vectors = [];
-        return false;
-      })
-      .sort((a, b) => a[1].updatedAt - b[1].updatedAt);
-    this.tokenBatchCheckpoints.clear();
-    for (const [key, checkpoint] of retained.slice(
-      -MAX_COMPLETED_EMBED_RESULTS,
-    )) {
-      this.tokenBatchCheckpoints.set(key, checkpoint);
-    }
-    for (const [, checkpoint] of retained.slice(
-      0,
-      -MAX_COMPLETED_EMBED_RESULTS,
-    )) {
-      checkpoint.vectors = [];
-    }
-  }
-
-  /** Consume one interrupted call's successful prefix. */
-  takeTokenBatchCheckpoint(key: string): TokenBatchCheckpoint | undefined {
-    this.pruneTokenBatchCheckpoints();
-    const checkpoint = this.tokenBatchCheckpoints.get(key);
-    if (!checkpoint) return undefined;
-    this.tokenBatchCheckpoints.delete(key);
-    return {
-      ...checkpoint,
-      vectors: cloneEmbeddingVectors(checkpoint.vectors),
-    };
-  }
-
-  /** Retain bounded progress only after an interrupted durable drain. */
-  storeTokenBatchCheckpoint(
-    key: string,
-    nextIndex: number,
-    vectors: Float32Array[],
-  ): void {
-    if (this.closing || nextIndex <= 0 || vectors.length !== nextIndex) return;
-    this.tokenBatchCheckpoints.set(key, {
-      nextIndex,
-      vectors: cloneEmbeddingVectors(vectors),
-      updatedAt: Date.now(),
-    });
-    this.pruneTokenBatchCheckpoints();
-  }
-
-  private recordSlotSuccess(slot: EmbedSlot): void {
-    slot.healthy = true;
-    if (
-      slot.recoveryGeneration !== null &&
-      slot.recoveryGeneration === localInitFailureGeneration
-    ) {
-      slot.recoveryGeneration = null;
-      if (localInitFailures > 0) {
-        log.info(
-          `local embedding provider recovered after ${localInitFailures} failed init attempt(s)`,
-        );
-      }
-      localInitFailures = 0;
-      localInitRetryAt = 0;
-      if (localProviderFailureCause === "transient-init-exhausted") {
-        clearLocalProviderLatch();
-      }
-      localProviderErrorLogged = false;
-      return;
-    }
-
-    slot.recoveryGeneration = null;
-    // An already in-flight sibling proves service is still available, but must
-    // not erase another slot's retry debt or admit immediate respawns.
-    this.preserveHealthyServiceAfterExhaustion();
-  }
-
-  private async runOperation(
-    operation: PoolOperation,
-    slot: EmbedSlot,
-  ): Promise<void> {
-    let watchdogTimer: ReturnType<typeof setTimeout> | undefined;
-    let rejectWatchdog!: (error: Error) => void;
-    const watchdog = new Promise<never>((_resolve, reject) => {
-      rejectWatchdog = reject;
-    });
-    const armWatchdog = (
-      stage: "init" | "execution",
-      timeoutMs: number,
-    ): void => {
-      if (watchdogTimer) clearTimeout(watchdogTimer);
-      watchdogTimer = setTimeout(
-        () => {
-          log.error(
-            `embedding worker watchdog expired: stage=${stage} timeout_ms=${timeoutMs}`,
-          );
-          rejectWatchdog(new EmbeddingWorkerWatchdogError(stage));
-        },
-        Math.max(1, timeoutMs),
-      );
-      watchdogTimer.unref?.();
-    };
-    armWatchdog("init", embedInitWatchdogMs);
-
-    try {
-      // Deliberately omit every caller's AbortSignal. The operation owns the
-      // worker slot until native inference settles; callers only own waiters.
-      const vectors = await Promise.race([
-        slot.provider.embed(
-          operation.texts,
-          operation.inputType,
-          undefined,
-          () => armWatchdog("execution", embedExecutionWatchdogMs),
-        ),
-        watchdog,
-      ]);
-      this.recordSlotSuccess(slot);
-      const retainResult =
-        operation.retainResult === true && operation.waiters.size === 0;
-      operation.state = "completed";
-      operation.completedAt = Date.now();
-      operation.vectors = retainResult
-        ? cloneEmbeddingVectors(vectors)
-        : undefined;
-      operation.texts = [];
-      for (const waiter of operation.waiters) {
-        this.settleWaiter(operation, waiter, { vectors });
-      }
-      if (retainResult) {
-        this.pruneCompletedOperations(operation.completedAt);
-      } else if (this.operations.get(operation.key) === operation) {
-        this.operations.delete(operation.key);
-      }
-    } catch (error) {
-      if (this.operations.get(operation.key) === operation) {
-        this.operations.delete(operation.key);
-      }
-      operation.texts = [];
-      operation.vectors = undefined;
-      const ownedError =
-        error instanceof Error
-          ? error
-          : new Error("embedding worker operation failed");
-      for (const waiter of operation.waiters) {
-        this.settleWaiter(operation, waiter, { error: ownedError });
-      }
-      if (
-        error instanceof LocalProviderUnavailableError ||
-        error instanceof EmbeddingWorkerWatchdogError
-      ) {
-        await this.retireSlot(slot);
-      }
-    } finally {
-      if (watchdogTimer) clearTimeout(watchdogTimer);
-      slot.inflight--;
-      this.dispatch();
-    }
-  }
-
-  private dispatch(): void {
-    if (this.closing || this.dispatching) return;
-    this.dispatching = true;
-    try {
-      while (this.queue.length > 0) {
-        const operation = this.queue[0];
-        if (operation.waiters.size === 0) {
-          this.dropUnobservedQueuedOperation(operation);
-          continue;
-        }
-
-        // A retiring worker still owns its native model/heap until confirmed
-        // exit. Do not replace the last active slot (or exceed the ceiling via
-        // pool growth) while that memory remains live. Retirement completion
-        // re-enters dispatch above.
-        if (this.slots.length === 0 && this.retiredShutdowns.size > 0) break;
-
-        let slot: EmbedSlot;
-        try {
-          slot = this.pickSlot();
-        } catch (error) {
-          if (error instanceof EmbeddingWorkerRetryCooldownError) {
-            this.scheduleRetryDispatch(error.retryAt);
-            break;
-          }
-          this.removeQueuedOperation(0);
-          if (this.operations.get(operation.key) === operation) {
-            this.operations.delete(operation.key);
-          }
-          const ownedError =
-            error instanceof Error
-              ? error
-              : new LocalProviderUnavailableError();
-          for (const waiter of operation.waiters) {
-            this.settleWaiter(operation, waiter, { error: ownedError });
-          }
-          continue;
-        }
-        this.clearRetryDispatchTimer();
-        if (slot.inflight > 0) break;
-
-        this.removeQueuedOperation(0);
-        operation.state = "running";
-        slot.inflight++;
-        void this.runOperation(operation, slot);
-      }
-    } finally {
-      this.dispatching = false;
-    }
-  }
-
-  embed(
-    texts: string[],
-    inputType: "document" | "query",
-    signal?: AbortSignal,
-  ): Promise<Float32Array[]> {
-    if (signal?.aborted) {
-      return Promise.reject(new EmbeddingRequestAbortedError());
-    }
-    if (this.closing) {
-      return Promise.reject(
-        new LocalProviderUnavailableError("embedding pool is unavailable"),
-      );
-    }
-
-    this.pruneCompletedOperations();
-    const ownedTexts = texts.slice();
-    const key = embeddingOperationKey(ownedTexts, inputType);
-    const existing = this.operations.get(key);
-    if (existing) return this.attachWaiter(existing, signal);
-    const priority = isRecallEmbed(ownedTexts, inputType) ? "high" : "normal";
-    const byteSize = ownedTexts.reduce(
-      (total, text) => total + Buffer.byteLength(text),
-      0,
-    );
-    if (!this.canEnqueueOperation(priority, byteSize)) {
-      return Promise.reject(new EmbeddingQueueCapacityError());
-    }
-
-    const operation: PoolOperation = {
-      key,
-      texts: ownedTexts,
-      byteSize,
-      inputType,
-      priority,
-      state: "queued",
-      waiters: new Set(),
-    };
-    this.operations.set(key, operation);
-    const result = this.attachWaiter(operation, signal);
-    if (operation.waiters.size > 0) {
-      this.enqueueOperation(operation);
-      this.dispatch();
-    }
-    return result;
-  }
-
-  hasHealthySlot(): boolean {
-    return this.slots.some((slot) => slot.healthy);
-  }
-
-  /** Shut every worker down and clear the pool. Resolves once all have exited
-   *  (or been force-terminated). Idempotent. */
-  shutdown(timeoutMs = WORKER_SHUTDOWN_TIMEOUT_MS): Promise<void> {
-    if (this.shutdownPromise) return this.shutdownPromise;
-    this.closing = true;
-    this.clearRetryDispatchTimer();
-    const shutdownError = new LocalProviderUnavailableError(
-      "embedding pool shut down",
-    );
-    for (const operation of this.operations.values()) {
-      for (const waiter of operation.waiters) {
-        this.settleWaiter(operation, waiter, { error: shutdownError });
-      }
-      operation.texts = [];
-      operation.vectors = undefined;
-    }
-    this.queue.length = 0;
-    this.queuedBytes = 0;
-    this.operations.clear();
-    for (const checkpoint of this.tokenBatchCheckpoints.values()) {
-      checkpoint.vectors = [];
-    }
-    this.tokenBatchCheckpoints.clear();
-    const providers = this.slots.splice(0).map((s) => s.provider);
-    this.shutdownPromise = (async () => {
-      const recordedRetirementFailures = this.retiredShutdownErrors.length;
-      const failures = [...this.retiredShutdownErrors];
-      const active = await Promise.allSettled(
-        providers.map((provider) => provider.shutdown(timeoutMs)),
-      );
-      for (const outcome of active) {
-        if (outcome.status === "rejected") failures.push(outcome.reason);
-      }
-      // Retirements can be added by request rejection microtasks triggered by
-      // the active shutdowns above, so drain until the owned set stays empty.
-      while (this.retiredShutdowns.size > 0) {
-        await Promise.all(this.retiredShutdowns);
-      }
-      failures.push(
-        ...this.retiredShutdownErrors.slice(recordedRetirementFailures),
-      );
-      if (failures.length > 0) {
-        throw new AggregateError(
-          failures,
-          "embedding worker termination was not confirmed",
-        );
-      }
-    })();
-    return this.shutdownPromise;
-  }
-}
-
-/** Minimal worker surface needed to shut a worker down — lets tests inject a
- *  fake worker without spawning a real thread. */
-export interface ShutdownableWorker {
-  on(event: "exit", listener: () => void): unknown;
-  postMessage(value: WorkerInbound): void;
-  terminate(): Promise<number>;
-}
-
-/**
- * Ask a worker to exit cooperatively, but never wait longer than `timeoutMs`:
- * on timeout, force-`terminate()` it. Resolves only once the worker has exited
- * or `terminate()` confirms its exit; rejects if termination cannot be
- * confirmed.
- *
- * Exported (underscore-free name is fine; it's a real helper) so the bounded
- * shutdown can be unit-tested with a fake worker — the real failure mode is a
- * worker stuck in an uninterruptible single-threaded ONNX inference batch that
- * never emits "exit", which would otherwise hang process shutdown.
- */
-export function awaitWorkerShutdown(
-  worker: ShutdownableWorker,
-  timeoutMs: number,
-): Promise<void> {
-  return new Promise<void>((resolve, reject) => {
-    let done = false;
-    let forceStarted = false;
-    const finish = () => {
-      if (done) return;
-      done = true;
-      clearTimeout(killTimer);
-      resolve();
-    };
-    const forceTerminate = (): void => {
-      if (done || forceStarted) return;
-      forceStarted = true;
-      void worker.terminate().then(finish, (cause: unknown) => {
-        if (done) return;
-        done = true;
-        clearTimeout(killTimer);
-        reject(
-          new Error("embedding worker termination was not confirmed", {
-            cause,
-          }),
-        );
-      });
-    };
-    // Hard cap: if the worker is mid-inference (an uninterruptible
-    // single-threaded ONNX batch) and never emits "exit", force-terminate it
-    // so process shutdown can't hang. Terminating is safe — all SQLite state
-    // lives on the main thread; the worker is stateless.
-    const killTimer = setTimeout(forceTerminate, Math.max(0, timeoutMs));
-    killTimer.unref?.();
-
-    worker.on("exit", finish);
-    try {
-      worker.postMessage({ type: "shutdown" } satisfies WorkerInbound);
-    } catch {
-      // Posting can also fail during a termination race. Ask Node to terminate
-      // and use that promise as the authoritative exit confirmation.
-      forceTerminate();
-    }
-  });
-}
-
-// ---------------------------------------------------------------------------
-// Provider resolution
-// ---------------------------------------------------------------------------
-
-/** Default models per provider — used when config doesn't override. */
-const PROVIDER_DEFAULTS: Record<string, { model: string; dimensions: number }> =
-  {
-    local: { model: "nomic-ai/nomic-embed-text-v1.5", dimensions: 768 },
-    voyage: { model: "voyage-code-3", dimensions: 1024 },
-    openai: { model: "text-embedding-3-small", dimensions: 1536 },
-  };
-
-/** Env var name for each provider's API key. */
-const PROVIDER_ENV_KEYS: Record<string, string> = {
-  voyage: "VOYAGE_API_KEY",
-  openai: "OPENAI_API_KEY",
-};
-
-function getProviderApiKey(provider: string): string | undefined {
-  const envKey = PROVIDER_ENV_KEYS[provider];
-  return envKey ? process.env[envKey] : undefined;
-}
-
-let cachedProvider: EmbeddingProvider | null | undefined;
-/** One-way process-shutdown latch. Config resets must never reopen admission. */
-let providerAdmissionClosed = false;
-/** Every local-provider generation remains owned until worker exit is known. */
-const providerGenerationShutdowns = new Set<Promise<void>>();
-/** An unconfirmed exit is permanent for this process generation. */
-const providerGenerationShutdownErrors: unknown[] = [];
-
-function trackProviderGenerationShutdown(
-  shutdown: Promise<void>,
-): Promise<void> {
-  let tracked: Promise<void>;
-  tracked = shutdown
-    .then(
-      () => {},
-      (error: unknown) => {
-        providerGenerationShutdownErrors.push(error);
-      },
-    )
-    .finally(() => providerGenerationShutdowns.delete(tracked));
-  providerGenerationShutdowns.add(tracked);
-  return shutdown;
-}
-
-async function settleProviderGenerations(): Promise<void> {
-  // Completion callbacks can add a retirement while an active generation is
-  // settling, so drain until ownership stays empty.
-  while (providerGenerationShutdowns.size > 0) {
-    await Promise.all(providerGenerationShutdowns);
-  }
-  if (providerGenerationShutdownErrors.length > 0) {
-    throw new AggregateError(
-      providerGenerationShutdownErrors,
-      "embedding worker termination was not confirmed",
-    );
-  }
-}
-
-function getProvider(): EmbeddingProvider | null {
-  if (providerAdmissionClosed) return null;
-  if (cachedProvider !== undefined) return cachedProvider;
-
-  const cfg = config().search.embeddings;
-  if (cfg.enabled === false) {
-    cachedProvider = null;
-    return null;
-  }
-
-  const providerName = cfg.provider;
-  const model = cfg.model;
-
-  switch (providerName) {
-    case "local": {
-      // Construct the provider optimistically — the ONNX model init
-      // happens lazily in the worker thread(s) on first `embed()` call.
-      // If it fails, `LocalProviderUnavailableError` marks the provider
-      // as broken and callers degrade to FTS-only search. The pool wraps
-      // one or more LocalProvider workers (memory-gated) so concurrent
-      // embeds run in parallel instead of serializing (#999).
-      cachedProvider = new EmbeddingPool(model, cfg.dimensions);
-      break;
-    }
-    case "voyage": {
-      const apiKey = getProviderApiKey(providerName);
-      if (!apiKey) {
-        cachedProvider = null;
-        return null;
-      }
-      cachedProvider = new VoyageProvider(apiKey, model, cfg.dimensions);
-      break;
-    }
-    case "openai": {
-      const apiKey = getProviderApiKey(providerName);
-      if (!apiKey) {
-        cachedProvider = null;
-        return null;
-      }
-      cachedProvider = new OpenAIProvider(apiKey, model, cfg.dimensions);
-      break;
-    }
-    default:
-      log.info(`unknown embedding provider: ${String(providerName)}`);
-      cachedProvider = null;
-  }
-
-  return cachedProvider;
-}
-
-/** Reset cached provider — called when config changes.
- *  Shuts down the worker thread(s) if the current provider is a local pool.
- *  Returns a promise that resolves once all workers have fully exited.
- *  Callers that need clean teardown (tests) should await the result. */
-export function resetProvider(timeoutMs?: number): Promise<void> {
-  let shutdownPromise: Promise<void> = Promise.resolve();
-  if (cachedProvider instanceof EmbeddingPool) {
-    shutdownPromise = trackProviderGenerationShutdown(
-      cachedProvider.shutdown(timeoutMs),
-    );
-  }
-  // A config reset normally admits a fresh generation. Once process shutdown
-  // owns teardown, preserve the closed latch so a late callback cannot reopen
-  // the pool after shutdownProvider has snapshotted all owned generations.
-  cachedProvider = providerAdmissionClosed ? null : undefined;
-  return shutdownPromise;
-}
-
-/** Shut down the current provider and prevent any new provider from being
- *  created. After this call, `embed()` throws and `isAvailable()` returns
- *  false. Waits for every previous provider generation too, so a config reset
- *  racing process shutdown cannot orphan its old worker. */
-export async function shutdownProvider(timeoutMs?: number): Promise<void> {
-  providerAdmissionClosed = true;
-  if (cachedProvider instanceof EmbeddingPool) {
-    void trackProviderGenerationShutdown(cachedProvider.shutdown(timeoutMs));
-  }
-  cachedProvider = null; // null (not undefined) → getProvider() returns null, won't create new
-  await settleProviderGenerations();
-}
-
-/** @deprecated Test compatibility alias; use {@link shutdownProvider}. */
-export const _shutdownAndDisable = shutdownProvider;
-
-/** Test-only: clear settled generation failures between isolated cases. */
-export function _resetProviderShutdownTrackingForTest(): void {
-  if (providerGenerationShutdowns.size > 0) {
-    throw new Error("cannot reset embedding shutdown tracking while active");
-  }
-  providerGenerationShutdownErrors.length = 0;
-}
-
-/** Save the current cached provider reference (including the live worker)
- *  and clear the cache so the next `getProvider()` call creates a fresh one.
- *  Returns an opaque token that must be passed to `_restoreProvider()` to
- *  put the original provider back — without this, the worker is orphaned and
- *  a second ONNX load in the same Bun process will crash.
- *
- *  Test-only helper: lets suites temporarily swap in a mock/unavailable
- *  provider without killing the real worker. */
-export function _saveAndClearProvider(): unknown {
-  const saved = {
-    provider: cachedProvider,
-    admissionClosed: providerAdmissionClosed,
-  };
-  providerAdmissionClosed = false;
-  cachedProvider = undefined;
-  return saved;
-}
-
-/** Restore a provider previously saved by `_saveAndClearProvider()`. Any
- *  provider created between save and restore is discarded (callers must
- *  ensure it's not a LocalProvider with a live worker — those suites only
- *  use `_markLocalProviderUnavailable()` so no worker is spawned). */
-export function _restoreProvider(token: unknown): void {
-  const saved = token as {
-    provider: EmbeddingProvider | null | undefined;
-    admissionClosed?: boolean;
-  };
-  providerAdmissionClosed = saved.admissionClosed ?? false;
-  cachedProvider = saved.provider;
-}
-
-/**
- * Quick sanity check that a string looks like a real API key rather than
- * a placeholder. Tools like Codex set `OPENAI_API_KEY=nokey` when routing
- * through a custom base URL — using such a value for real API calls
- * produces 401 errors.
- */
-function looksLikeApiKey(key: string): boolean {
-  // Real API keys are at least 20 characters, don't contain whitespace,
-  // and aren't common placeholders.
-  const trimmed = key.trim();
-  return trimmed.length >= 20 && !/\s/.test(trimmed);
-}
-
-/**
- * Build a remote `EmbeddingProvider` from whichever API key is in env.
- * Returns `null` when neither `VOYAGE_API_KEY` nor `OPENAI_API_KEY` is set,
- * which is the signal for callers to fall through to FTS-only behaviour.
- *
- * Voyage wins ties because it's the higher-quality option for code search;
- * users who want OpenAI specifically can pin `search.embeddings.provider`
- * in `.lore.json` and skip the fallback path entirely.
- */
-export function pickRemoteFallback(): {
-  name: "voyage" | "openai";
-  provider: EmbeddingProvider;
-} | null {
-  // Validate keys before using them — tools like Codex/OpenCode often set
-  // OPENAI_API_KEY to a placeholder (e.g. "nokey") when using a custom
-  // OPENAI_BASE_URL. Using such a key for real API calls produces 401 noise.
-  const voyageKey = process.env.VOYAGE_API_KEY;
-  if (voyageKey && looksLikeApiKey(voyageKey)) {
-    const d = PROVIDER_DEFAULTS.voyage;
-    return {
-      name: "voyage",
-      provider: new VoyageProvider(voyageKey, d.model, d.dimensions),
-    };
-  }
-  const openaiKey = process.env.OPENAI_API_KEY;
-  if (openaiKey && looksLikeApiKey(openaiKey)) {
-    const d = PROVIDER_DEFAULTS.openai;
-    return {
-      name: "openai",
-      provider: new OpenAIProvider(openaiKey, d.model, d.dimensions),
-    };
-  }
-  return null;
-}
-
-// ---------------------------------------------------------------------------
-// Availability
-// ---------------------------------------------------------------------------
-
-/** Returns true if embedding is available.
- *  Active when the configured provider's API key is set, unless explicitly
- *  disabled via `search.embeddings.enabled: false` in .lore.json.
- *
- *  For the `local` provider, also returns false once the worker has reported
- *  an init failure — callers (recall, ltm, distillation) use this gate to
- *  skip embedding work and fall back to FTS-only search. */
-export function isAvailable(): boolean {
-  const provider = getProvider();
-  if (!provider) return false;
-  if (provider instanceof EmbeddingPool) {
-    if (
-      localProviderKnownUnavailable() &&
-      !(
-        localProviderFailureCause === "transient-init-exhausted" &&
-        provider.hasHealthySlot()
-      )
-    ) {
-      // One-time log so the user knows why vector search is degraded.
-      if (!localProviderErrorLogged) {
-        localProviderErrorLogged = true;
-        log.info(
-          "local embedding provider unavailable — recall will use FTS-only search. " +
-            "To use a remote provider, set search.embeddings.provider in .lore.json.",
-        );
-      }
-      return false;
-    }
-    if (localInitRetryAt > 0) {
-      // A proven sibling remains usable while a failed slot cools down. The
-      // pool blocks replacement spawning but continues routing to that sibling.
-      if (provider.hasHealthySlot()) return true;
-      // A transient init failure is cooling down before its next retry.
-      if (Date.now() < localInitRetryAt) return false; // FTS-only until then
-      // Cooldown elapsed: report available so the next embed can enter the
-      // pool, which admits exactly one fresh recovery slot and clears the arm.
-    }
-  }
-  return true;
-}
-
-/** Coarse embedding-subsystem state for health surfaces. */
-export type EmbeddingHealthState =
-  | "ok"
-  | "retrying"
-  | "unavailable"
-  | "disabled";
-
-export interface EmbeddingHealth {
-  /** Whether vector recall is currently usable. */
-  available: boolean;
-  state: EmbeddingHealthState;
-  provider: "local" | "remote" | "none";
-  /** Human-readable explanation for logs / `lore doctor` / `/health`. */
-  detail: string;
-}
-
-/**
- * Read-only snapshot of embedding availability for health surfaces
- * (`/health`, `lore doctor`). Unlike {@link isAvailable}, this has NO side
- * effects: it never logs, never triggers a retry, and never spawns a worker.
- * When it reports `unavailable` or `retrying`, recall is running FTS-only —
- * the degradation that was previously visible only as a one-time log line.
- */
-export function embeddingStatus(): EmbeddingHealth {
-  const provider = getProvider();
-  if (!provider) {
-    return {
-      available: false,
-      state: "disabled",
-      provider: "none",
-      detail:
-        "no embedding provider configured (or disabled) — recall uses FTS-only search",
-    };
-  }
-  if (provider instanceof EmbeddingPool) {
-    if (
-      localProviderKnownUnavailable() &&
-      !(
-        localProviderFailureCause === "transient-init-exhausted" &&
-        provider.hasHealthySlot()
-      )
-    ) {
-      return {
-        available: false,
-        state: "unavailable",
-        provider: "local",
-        detail:
-          "local embedding provider failed to initialize — recall is FTS-only (semantic search degraded)",
-      };
-    }
-    if (localInitRetryAt > 0) {
-      if (provider.hasHealthySlot()) {
-        return {
-          available: true,
-          state: "ok",
-          provider: "local",
-          detail:
-            "local ONNX embeddings active on a healthy pool slot (failed slot retry pending)",
-        };
-      }
-      // A transient init failure has a retry armed. This is a pure read, so —
-      // unlike isAvailable() — we do NOT clear the deadline or trigger the
-      // reset here; the provider is not confirmed healthy until a later embed
-      // re-inits it. Report "retrying" (recall is FTS-only) for the whole
-      // armed window, whether or not the cooldown has elapsed, so health checks
-      // never show a false "ok" before recovery actually happens.
-      return {
-        available: false,
-        state: "retrying",
-        provider: "local",
-        detail:
-          "local embedding provider init failed; a retry is armed — recall is FTS-only until it recovers",
-      };
-    }
-    return {
-      available: true,
-      state: "ok",
-      provider: "local",
-      detail: "local ONNX embeddings active (vector recall enabled)",
-    };
-  }
-  return {
-    available: true,
-    state: "ok",
-    provider: "remote",
-    detail: "remote embedding provider active (vector recall enabled)",
-  };
-}
-
-async function waitForEmbeddingRetry(
-  delayMs: number,
-  guard: EmbeddingAbortGuard,
-): Promise<void> {
-  if (delayMs <= 0) {
-    throwIfEmbeddingAborted(guard);
-    return;
-  }
-  let timer: ReturnType<typeof setTimeout> | undefined;
-  const wait = new Promise<void>((resolve) => {
-    timer = setTimeout(resolve, delayMs);
-  });
-  try {
-    await awaitEmbeddingOperation(wait, guard);
-  } finally {
-    if (timer) clearTimeout(timer);
-  }
-}
-
-/**
- * Await a proven local embedding worker, including bounded transient-init
- * retries. Remote providers return immediately to avoid spending API quota.
- *
- * This is for orchestration that requires vectors (not normal recall, which may
- * degrade to FTS). The readiness probe serializes cold model download/load,
- * waits for the provider-owned 2s/8s retry deadlines, and stops on the caller's
- * cancellation boundary or a terminal/exhausted provider failure.
- */
-export async function ensureEmbeddingReady(
-  options: EmbeddingOperationOptions = {},
-): Promise<void> {
-  const guard = createEmbeddingAbortGuard("provider-readiness", options);
-  throwIfEmbeddingAborted(guard);
-
-  for (;;) {
-    throwIfEmbeddingAborted(guard);
-    const provider = getProvider();
-    if (!provider) {
-      throw new LocalProviderUnavailableError(
-        "no embedding provider is configured",
-      );
-    }
-    if (!(provider instanceof EmbeddingPool)) {
-      throwIfEmbeddingAborted(guard);
-      return;
-    }
-    if (
-      localProviderFailureCause === "terminal" ||
-      (localProviderFailureCause === "transient-init-exhausted" &&
-        !provider.hasHealthySlot())
-    ) {
-      throw new LocalProviderUnavailableError(
-        "local embedding provider is unavailable",
-      );
-    }
-    if (provider.hasHealthySlot()) {
-      throwIfEmbeddingAborted(guard);
-      return;
-    }
-
-    try {
-      await awaitEmbeddingOperation(
-        provider.embed(["warmup"], "document"),
-        guard,
-      );
-      throwIfEmbeddingAborted(guard);
-      return;
-    } catch (error) {
-      if (error instanceof EmbeddingAbortError) throw error;
-      // Worker watchdogs are an internal lifecycle classification. Keep the
-      // public readiness contract provider-shaped so orchestration that already
-      // handles an unavailable local provider does not need to know which
-      // worker-owned deadline expired.
-      if (error instanceof EmbeddingWorkerWatchdogError) {
-        throw new LocalProviderUnavailableError(error);
-      }
-      if (!(error instanceof LocalProviderUnavailableError)) throw error;
-      throwIfEmbeddingAborted(guard);
-      if (localProviderFailureCause !== null || localInitRetryAt <= 0) {
-        throw error;
-      }
-      await waitForEmbeddingRetry(
-        Math.max(0, localInitRetryAt - Date.now()),
-        guard,
-      );
-    }
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Public embed API
-// ---------------------------------------------------------------------------
-
-/**
- * A single-text `query` embed is a recall lookup — the latency-sensitive
- * request path (the recall tool and forSession LTM ranking), as opposed to
- * batch/document writes. This is the single source of truth for that
- * classification: the pool uses it to assign "high" priority and {@link embed}
- * uses it to track in-flight
- * recall load for the temporal backfill's idle-gate. Both sites call this one
- * predicate so the "mirrors the worker" invariant can never silently drift.
- */
-function isRecallEmbed(
-  texts: string[],
-  inputType: "document" | "query",
-): boolean {
-  return inputType === "query" && texts.length === 1;
-}
-
-/**
- * Number of recall (single query-text) embeds currently in flight through
- * {@link embed}. The temporal re-chunk backfill reads this (via
- * {@link recallEmbedsInFlight}) to yield the shared embedding worker to
- * latency-sensitive recall lookups: it parks a page while a recall embed is
- * outstanding and resumes the instant the worker drains.
- *
- * Only single-text `query` embeds are counted — that is exactly the class the
- * worker itself marks "high priority" (see LocalProvider.embed). Document/batch
- * embeds (write-time message/knowledge/entity embeds AND the backfill's own
- * embeds) are deliberately NOT counted, so the backfill never gates against its
- * own load or against fire-and-forget background writes.
- */
-let _recallEmbedsInFlight = 0;
-
-/**
- * Live count of in-flight recall (single query-text) embeds — the temporal
- * re-chunk backfill's idle signal. Read on every gate poll so it reflects the
- * shared worker's current recall load, never a stale snapshot.
- */
-export function recallEmbedsInFlight(): number {
-  return _recallEmbedsInFlight;
-}
-
-/**
- * Test seam: force the in-flight recall-embed counter to a known value so the
- * gateway's idle-gate wiring can be exercised deterministically without driving
- * a real (async, provider-dependent) embed. Test-only.
- */
-export function _setRecallEmbedsInFlightForTest(n: number): void {
-  _recallEmbedsInFlight = n;
-}
-
-/**
- * Generate embeddings for the given texts using the configured provider.
- *
- * Remote providers (voyage, openai) are explicit opt-in via
- * `search.embeddings.provider` in `.lore.json` — there is no automatic
- * fallback from local to remote. When the local provider is unavailable,
- * callers should gate on `isAvailable()` and degrade to FTS-only search.
- *
- * @param texts     Array of texts to embed
- * @param inputType "document" for storage, "query" for search
- * @returns         Float32Array per input text
- * @throws          On API errors or when no provider is available
- */
-export async function embed(
-  texts: string[],
-  inputType: "document" | "query",
-  signal?: AbortSignal,
-): Promise<Float32Array[]> {
-  const provider = getProvider();
-  if (!provider) throw new Error("No embedding provider available");
-  return await embedWithProvider(provider, texts, inputType, signal);
-}
-
-/** Run one request against a fixed provider instance. Keeping provider capture
- * explicit lets a multi-batch operation survive config/reset races without
- * silently moving later batches to a different model. */
-async function embedWithProvider(
-  provider: EmbeddingProvider,
-  texts: string[],
-  inputType: "document" | "query",
-  signal?: AbortSignal,
-): Promise<Float32Array[]> {
-  // A single-text query embed is a recall lookup (see isRecallEmbed — the same
-  // predicate the worker uses for high priority). Track it in flight so the
-  // temporal re-chunk backfill can yield the shared worker while recall is
-  // active. The counter is decremented in `finally` so a rejected embed
-  // (provider gone, OOM, timeout) can never leak a permanent "busy" that would
-  // wedge the backfill forever.
-  const isRecall = isRecallEmbed(texts, inputType);
-  if (isRecall) _recallEmbedsInFlight++;
-  try {
-    const vecs = await provider.embed(texts, inputType, signal);
-    // Enforce the L2-normalization invariant at the single chokepoint so the JS
-    // dot-product path and sqlite-vec's vec_distance_cosine() always agree. See
-    // l2Normalize() for the full rationale.
-    return vecs.map(l2Normalize);
-  } finally {
-    if (isRecall) _recallEmbedsInFlight--;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Cosine similarity (pure JS)
-// ---------------------------------------------------------------------------
-
-/**
- * Return an L2-normalized (unit-length) copy of a vector.
- *
- * System-wide invariant: every vector produced by {@link embed} is
- * L2-normalized. Two consumers depend on this and would silently DIVERGE if it
- * were violated:
- *   1. {@link cosineSimilarity} (the JS brute-force path) is a bare dot product
- *      — it only equals cosine similarity for unit vectors.
- *   2. sqlite-vec's `vec_distance_cosine()` (the native fast path) normalizes
- *      internally, so a non-normalized stored vector would score differently on
- *      the two paths (e.g. stored [2,0,0] vs query [1,0,0]: JS dot → 2.0, vec →
- *      1.0), breaking vec/JS parity and producing similarities outside [-1, 1].
- *
- * Providers (local ONNX, Voyage, OpenAI) already return ~unit vectors, so this
- * is idempotent in practice. It is applied unconditionally at the single
- * {@link embed} chokepoint to make the invariant true *by construction* rather
- * than by provider convention — guarding against drift if a provider ever
- * returns non-normalized output.
- *
- * A zero or non-finite vector cannot be normalized and is returned unchanged
- * (matches {@link cosineSimilarity}'s zero-vector handling, which returns 0).
- */
-export function l2Normalize(vec: Float32Array): Float32Array {
-  let sumSq = 0;
-  for (let i = 0; i < vec.length; i++) sumSq += vec[i] * vec[i];
-  const norm = Math.sqrt(sumSq);
-  if (!(norm > 0) || !Number.isFinite(norm)) return vec;
-  const out = new Float32Array(vec.length);
-  for (let i = 0; i < vec.length; i++) out[i] = vec[i] / norm;
-  return out;
-}
-
-// ---------------------------------------------------------------------------
-// Vector search — knowledge
-// ---------------------------------------------------------------------------
-
-/**
- * Search all knowledge entries with embeddings by cosine similarity.
- * Returns top-k entries sorted by similarity descending.
- *
- * Uses dot-product (vectors are L2-normalized) and bounded top-k insertion
- * to avoid a full O(n log n) sort.
- *
- * @param excludeCategories  Optional category names to exclude from results.
- *   Useful when preferences are injected in a separate system block and
- *   shouldn't compete for vector search slots with context-bound entries.
- */
-/**
- * Run `spec` on the read-worker pool when it's enabled and healthy (off the
- * main event loop), otherwise synchronously in-process. The pool call never
- * throws: it returns null when unavailable (disabled/broken/errored) so we
- * transparently fall back in-process, or the VECTOR_SEARCH_TIMED_OUT sentinel
- * when the worker is alive but slow — in which case we return an empty result
- * rather than re-running the scan on the main thread.
- */
-async function poolOrInProcess(
-  spec: VectorQuerySpec,
-  queryEmbedding: Float32Array,
-): Promise<VectorHit[] | DistillationVectorHit[]> {
-  // #1065: record the wall-clock latency of every vector KNN read (pool
-  // round-trip + IPC, or the in-process fallback scan) tagged by the DB's
-  // (storage layout × sqlite-vec availability) cohort, so the gateway can prove
-  // the vec0 latency win (p50/p95) and spot a silently degraded JS-fallback
-  // host. storage_mode is DB-global and vec availability is process-global (the
-  // main thread and the workers load the extension together), so this
-  // main-thread resolution faithfully labels the worker path too. The lookup is
-  // a single indexed kv_meta point read (microseconds). Purely additive: the
-  // query path below is unchanged — the fallback re-resolves its own readMode.
-  const started = performance.now();
-  const cohort = resolveReadMode(readStorageMode(db()), isVecAvailable());
-  try {
-    const pooled = await tryPoolVectorSearch(spec, queryEmbedding);
-    // Timed out: the worker is alive but slow. Return empty — degrading this one
-    // recall — rather than re-running the O(n) scan on the main thread, which
-    // re-blocks the event loop (the stall bug). The pool cancels the timed-out
-    // worker query (terminates + respawns) so it recovers for the next caller.
-    if (pooled === VECTOR_SEARCH_TIMED_OUT) return [];
-    if (pooled !== null) return pooled;
-    const readMode = resolveReadMode(readStorageMode(db()), isVecAvailable());
-    const temporalPartitionMode =
-      spec.kind === "temporal" ? getKV(TEMPORAL_PARTITION_MODE_KEY) : null;
-    try {
-      return runVectorQuery(
-        db(),
-        readMode,
-        queryEmbedding,
-        spec,
-        temporalPartitionMode,
-      );
-    } catch (err) {
-      // Safety net for the vec0 read path, which (unlike the blob paths) has no
-      // in-line JS fallback: a vec0 `MATCH` can throw transiently — e.g. during a
-      // dimension change, when the query embedding's width no longer matches the
-      // vec0 table — and we must degrade THIS recall to empty (FTS/keyword recall
-      // still answers) rather than crash, per the never-crash contract. The pool
-      // already logged any worker-path failure; log the in-process one too so a
-      // systematic break stays visible.
-      log.error("in-process vector search failed; returning empty:", err);
-      return [];
-    }
-  } finally {
-    recordVecReadLatency(cohort, performance.now() - started);
-  }
-}
-
-export async function vectorSearch(
-  queryEmbedding: Float32Array,
-  limit = 10,
-  excludeCategories?: string[],
-): Promise<VectorHit[]> {
-  return await poolOrInProcess(
-    {
-      kind: "knowledge",
-      tenantId: currentTenantId(),
-      limit,
-      excludeCategories,
-    },
-    queryEmbedding,
-  );
-}
-
-/**
- * Search all entities with embeddings by cosine similarity.
- * Returns top-k entities sorted by similarity descending.
- */
-export async function vectorSearchEntities(
-  queryEmbedding: Float32Array,
-  limit = 10,
-): Promise<VectorHit[]> {
-  return await poolOrInProcess(
-    { kind: "entities", tenantId: currentTenantId(), limit },
-    queryEmbedding,
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Vector search — distillations
-// ---------------------------------------------------------------------------
-
-/**
- * Search non-archived distillations with embeddings by cosine similarity.
- * Returns top-k entries sorted by similarity descending.
- */
-export async function vectorSearchDistillations(
-  queryEmbedding: Float32Array,
-  limit = 10,
-): Promise<VectorHit[]> {
-  return await poolOrInProcess(
-    { kind: "distillations", tenantId: currentTenantId(), limit },
-    queryEmbedding,
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Vector search — all distillations (including archived)
-// ---------------------------------------------------------------------------
-
-/**
- * Search ALL distillations (including archived) with embeddings by cosine
- * similarity, scoped to a single project. Returns session_id alongside
- * similarity for cross-session counting.
- *
- * Unlike vectorSearchDistillations() which filters to non-archived only,
- * this searches the full distillation archive — necessary for detecting
- * repeated instructions across sessions where older distillations have
- * been archived after meta-distillation.
- *
- * Pure brute-force — fine for ~200 entries per project. Safety-capped
- * at 500 rows to prevent excessive CPU on long-running projects.
- */
-export async function vectorSearchAllDistillations(
-  queryEmbedding: Float32Array,
-  projectId: string,
-  limit = 20,
-): Promise<DistillationVectorHit[]> {
-  return (await poolOrInProcess(
-    { kind: "allDistillations", projectId, limit },
-    queryEmbedding,
-  )) as DistillationVectorHit[];
-}
-
-// ---------------------------------------------------------------------------
-// Fire-and-forget embedding
-// ---------------------------------------------------------------------------
-
-/**
- * Embed a knowledge entry and store the result in the DB.
- * Fire-and-forget — errors are logged, never thrown.
- * The entry remains usable via FTS even if embedding fails.
- */
-// Fire-and-forget document embeds (knowledge, distillation, entity) must not
-// block the write path, so callers don't await them. They are tracked here so
-// callers can drain them: at the test boundary (a promise that resolves AFTER
-// the harness closes/swaps the DB would otherwise write to the wrong DB or log
-// spurious errors — issue #885), and on graceful gateway shutdown (a
-// distillation embed still mid-flight when the worker is torn down never writes
-// its `distillation_vec` row → silent recall degradation on short/fast
-// sessions — issue #1331).
-const _docEmbedsInFlight = new Set<Promise<unknown>>();
-function trackDocEmbed(p: Promise<unknown>): void {
-  _docEmbedsInFlight.add(p);
-  void p.finally(() => _docEmbedsInFlight.delete(p));
-}
-
-function isExpectedBestEffortEmbeddingError(error: unknown): boolean {
-  return (
-    error instanceof LocalProviderUnavailableError ||
-    error instanceof EmbeddingQueueCapacityError
-  );
-}
-
-/**
- * Await all in-flight fire-and-forget document embeds (knowledge / distillation
- * / entity). Loops so embeds spawned while draining (rare) are also awaited.
- *
- * A numeric timeout preserves the shutdown contract: the drain resolves when
- * that best-effort deadline elapses even if embeds remain pending. The options
- * form accepts a caller signal and/or deadline and throws
- * {@link EmbeddingAbortError}, allowing semantic lint to report the interrupted
- * phase. Tests call it with no argument for a full (unbounded) drain.
- */
-export async function settleDocumentEmbeds(
-  timeoutOrOptions?: number | EmbeddingOperationOptions,
-): Promise<void> {
-  const legacyBestEffort = typeof timeoutOrOptions === "number";
-  const options: EmbeddingOperationOptions = legacyBestEffort
-    ? {
-        deadlineMs: Number.isFinite(timeoutOrOptions)
-          ? Math.max(0, timeoutOrOptions)
-          : 0,
-      }
-    : (timeoutOrOptions ?? {});
-  const guard = createEmbeddingAbortGuard("settle-document-embeds", options);
-
-  try {
-    while (_docEmbedsInFlight.size > 0) {
-      throwIfEmbeddingAborted(guard);
-      // Loop so embeds spawned mid-drain are picked up. An opted-in signal or
-      // deadline stops waiting promptly; already-running embeds remain tracked
-      // and can be drained by a later shutdown call.
-      await awaitEmbeddingOperation(
-        Promise.allSettled(_docEmbedsInFlight).then(() => undefined),
-        guard,
-      );
-    }
-    throwIfEmbeddingAborted(guard);
-  } catch (error) {
-    // Preserve the historical shutdown contract: the numeric form is a
-    // best-effort bounded drain that resolves at timeout. The options form is
-    // the typed lint/orchestration API and throws on abort/deadline.
-    if (
-      legacyBestEffort &&
-      error instanceof EmbeddingAbortError &&
-      error.code === "deadline-exceeded"
-    ) {
-      return;
-    }
-    throw error;
-  }
-}
-
-export function embedKnowledgeEntry(
-  id: string,
-  title: string,
-  content: string,
-): void {
-  if (!isAvailable()) return;
-  const text = `${title}\n${content}`;
-  trackDocEmbed(
-    embed([text], "document")
-      .then(([vec]) => {
-        storeEmbedding(db(), "knowledge", id, vec);
-      })
-      .catch((err) => {
-        if (isExpectedBestEffortEmbeddingError(err)) return;
-        log.error("embedding failed for knowledge entry", id, ":", err);
-      }),
-  );
-}
-
-/**
- * Embed an entity (canonical name + all alias values) and store the result.
- * Fire-and-forget — errors are logged, never thrown. Used by entity auto-dedup
- * (#462) to compute semantic similarity between entities whose names differ but
- * mean the same thing ("GitHub Actions" ↔ "GHA").
- */
-export function embedEntity(
-  id: string,
-  canonicalName: string,
-  aliasValues: string[],
-): void {
-  if (!isAvailable()) return;
-  // Composite text: canonical name followed by every alias value. Deduped and
-  // joined so the vector captures all surface forms of the entity.
-  const seen = new Set<string>();
-  const parts: string[] = [];
-  for (const v of [canonicalName, ...aliasValues]) {
-    const t = v.trim();
-    if (!t) continue;
-    const key = t.toLowerCase();
-    if (seen.has(key)) continue;
-    seen.add(key);
-    parts.push(t);
-  }
-  const text = parts.join(" ");
-  if (!text) return;
-  trackDocEmbed(
-    embed([text], "document")
-      .then(([vec]) => {
-        storeEmbedding(db(), "entities", id, vec);
-      })
-      .catch((err) => {
-        if (isExpectedBestEffortEmbeddingError(err)) return;
-        log.error("embedding failed for entity", id, ":", err);
-      }),
-  );
-}
-
-/**
- * Embed a distillation and store the result in the DB.
- * Fire-and-forget — errors are logged, never thrown.
- * The distillation remains searchable via FTS even if embedding fails.
- */
-export function embedDistillation(id: string, observations: string): void {
-  if (!isAvailable()) return;
-  trackDocEmbed(
-    embed([observations], "document")
-      .then(([vec]) => {
-        storeEmbedding(db(), "distillations", id, vec);
-      })
-      .catch((err) => {
-        if (isExpectedBestEffortEmbeddingError(err)) return;
-        log.error("embedding failed for distillation", id, ":", err);
-      }),
-  );
-}
-
-/**
- * Pay the local ONNX model's cold-load cost (~21s measured) up front by firing
- * one throwaway embed at gateway startup, so the FIRST real distillation embed
- * of the session is already warm and finishes within the turn instead of racing
- * teardown (#1331). Fire-and-forget: never blocks startup, never throws.
- *
- * Local-only: remote providers (Voyage/OpenAI) have no cold-load cost and a
- * warmup would only burn API quota, so this is a no-op for them and when
- * embeddings are disabled. Uses `"document"` inputType so it is NOT counted as
- * a recall embed (see {@link embed}) and cannot self-gate the temporal backfill.
- * Not tracked via `trackDocEmbed` — it stores nothing, so there is no
- * write-to-wrong-DB / mid-flight-on-teardown concern to drain.
- */
-export function warmupEmbedding(): void {
-  if (!isAvailable()) return;
-  if (config().search.embeddings.provider !== "local") return;
-  void embed(["warmup"], "document").catch((err) => {
-    if (isExpectedBestEffortEmbeddingError(err)) return;
-    log.error("embedding warmup failed:", err);
-  });
-}
-
-/**
- * Hard cap on how many vec0 chunks a single temporal message may fan out to.
- * A normal turn has a handful of parts; this only bites a pathological message
- * (e.g. hundreds of parallel tool calls). Past the cap the overflow units are
- * folded into ONE final chunk, so the per-message chunk count — which is both
- * `temporal_vec` rows AND the size of the KNN window `runTemporal` must collapse
- * by max-sim (then widen-retry over) — stays bounded, without silently dropping
- * any unit's text from the vector (FTS indexes the full content regardless).
- */
-export { MAX_TEMPORAL_CHUNKS_PER_MESSAGE } from "./embedding-units";
-
-/**
- * Embed `texts` in token-area-bounded sub-batches (via {@link nextBatch}) and
- * return the vectors in input order. The multi-vector temporal write hands us
- * one text per part-aware unit; a single `embed()` posts the WHOLE array to the
- * worker as one padded tensor, so a message with many large prose/reasoning
- * units could OOM-thrash the worker (the ×0.7 backoff lowers per-input length,
- * not input count). Sub-batching keeps each worker request's peak tensor within
- * MAX_BATCH_TOKEN_AREA, mirroring the backfill paths. All-or-nothing: a failure
- * in any batch rejects before the caller stores, so a partial chunk set is never
- * written (storeTemporalChunks DELETEs-then-INSERTs the complete set).
- *
- * A signal-driven local call checkpoints only its successfully completed prefix
- * when interrupted. Queued/current native work retains the pool's normal
- * cancellation semantics; no future sub-batch starts for the abandoned caller.
- * An identical durable retry on the same pool consumes the bounded checkpoint
- * once and resumes from that prefix instead of repeating successful inference.
- */
-export async function embedInTokenBatches(
-  texts: string[],
-  inputType: "document" | "query",
-  signal?: AbortSignal,
-): Promise<Float32Array[]> {
-  const provider = getProvider();
-  if (!provider) throw new Error("No embedding provider available");
-  const localPool =
-    provider instanceof EmbeddingPool && signal ? provider : undefined;
-  const checkpointKey = localPool
-    ? embeddingOperationKey(texts, inputType)
-    : undefined;
-  const checkpoint =
-    localPool && checkpointKey
-      ? localPool.takeTokenBatchCheckpoint(checkpointKey)
-      : undefined;
-  const items = texts.map((text) => ({ text }));
-  const out = checkpoint?.vectors ?? [];
-  let nextIndex = checkpoint?.nextIndex ?? 0;
-  try {
-    while (nextIndex < items.length) {
-      const batch = nextBatch(items, nextIndex);
-      const vecs = await embedWithProvider(
-        provider,
-        batch.map((b) => b.text),
-        inputType,
-        signal,
-      );
-      if (vecs.length !== batch.length) {
-        throw new Error(
-          "embedding provider returned an unexpected vector count",
-        );
-      }
-      out.push(...vecs);
-      nextIndex += batch.length;
-    }
-    return out;
-  } catch (error) {
-    if (localPool && checkpointKey && nextIndex > 0) {
-      localPool.storeTokenBatchCheckpoint(checkpointKey, nextIndex, out);
-    }
-    throw error;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Vector search — temporal messages (all, including distilled)
-// ---------------------------------------------------------------------------
-
-/**
- * Search temporal messages with embeddings by cosine similarity.
- * Returns top-k entries sorted by similarity descending.
- *
- * Includes distilled messages — their embeddings are preserved by
- * markDistilled() specifically to keep this search path viable.
- * Distilled messages contain specific details (algorithm names, config
- * values, file paths) that the distillation summary may have dropped.
- *
- * Scoped to a single project. Optionally scoped to a single session.
- */
-export async function vectorSearchTemporal(
-  queryEmbedding: Float32Array,
-  projectId: string,
-  limit = 10,
-  sessionId?: string,
-): Promise<VectorHit[]> {
-  return await poolOrInProcess(
-    { kind: "temporal", projectId, limit, sessionId },
-    queryEmbedding,
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Config change detection
-// ---------------------------------------------------------------------------
-
-/**
- * Build a config fingerprint from provider + model + dimensions.
- * Used to detect when the embedding config changes (provider swap, model swap,
- * dimension change) so we can clear stale embeddings and re-embed.
- *
- * 🟡 Scope note for FUTURE chunk/text-policy migrations: this fingerprint keys
- * ONLY on provider/model/dimensions — it deliberately does NOT include the
- * embedding TEXT policy (how a message is split into units / which parts are
- * embedded; see buildEmbeddingUnits). So changing that policy does NOT bump the
- * fingerprint, and the stale-embedding clear + backfill triggered here will NOT
- * fire for a pure policy change. Today the temporal re-chunk backfill re-embeds
- * the whole corpus under a new policy only because it is armed by an explicit KV
- * done-flag (see TEMPORAL_RECHUNK_DONE_KEY / resetTemporalRechunkProgress) that
- * is reset on vec0 cutover — NOT by this fingerprint. If a future migration
- * changes the chunk/text policy again WITHOUT a storage cutover, re-arm the walk
- * deliberately — fold a policy-version token into the done-flag key, or call
- * resetTemporalRechunkProgress() from the migration — because this fingerprint
- * will stay silent.
- */
-function configFingerprint(): string {
-  const cfg = config().search.embeddings;
-  return `${cfg.provider}:${cfg.model}:${cfg.dimensions}`;
-}
-
-const EMBEDDING_CONFIG_KEY = "lore:embedding_config";
-
-/**
- * Check if embedding config has changed since the last backfill.
- * If so, clear all existing embeddings (they're incompatible) and
- * update the stored fingerprint.
- *
- * Returns true if embeddings were cleared (full re-embed needed).
- */
-export function checkConfigChange(): boolean {
-  const current = configFingerprint();
-  const readStored = (): { value: string } | null =>
-    db()
-      .query("SELECT value FROM kv_meta WHERE key = ?")
-      .get(EMBEDDING_CONFIG_KEY) as { value: string } | null;
-  const observed = readStored();
-
-  if (observed?.value === current) return false;
-
-  const reconcile = (): boolean => {
-    // Re-check after BEGIN IMMEDIATE serializes competing processes. Without
-    // this, a late reconciler can clear vectors another process just rebuilt.
-    const stored = readStored();
-    if (stored?.value === current) return false;
-
-    const mode = readStorageMode(db());
-
-    // A vec0-store DB whose extension didn't load (degraded) cannot manage its
-    // embeddings — it can neither count/clear the unreadable vec0 tables nor
-    // recreate them. Leave the stored fingerprint UNCHANGED so the change is
-    // re-detected and handled the next time the DB opens on a capable runtime.
-    if (mode === "vec0" && !isVecAvailable()) return false;
-
-    // Config changed (or first run) — clear all embeddings in all tables
-    if (stored) {
-      const total =
-        mode === "vec0" ? countVec0Embeddings() : countBlobEmbeddings();
-      if (total > 0) {
-        clearAllEmbeddings(db());
-        log.info(
-          `embedding config changed (${stored.value} → ${current}), cleared ${total} stale embeddings`,
-        );
-      }
-      // A *dimension* change makes the fixed-width vec0 tables incompatible:
-      // recreate them at the new dimension (clearAllEmbeddings emptied the old
-      // rows; ensureVec0Store drops + recreates when the stored dim differs, and
-      // is a no-op for a same-dimension model/provider swap).
-      if (mode === "vec0") {
-        ensureVec0Store(db(), config().search.embeddings.dimensions);
-      }
-      // The clear wiped temporal vectors too, and temporal has no dedicated
-      // backfill loop above — re-arm the resumable re-chunk walk so it refills
-      // the corpus under the new model/dimension on this same startup.
-      resetTemporalRechunkProgress();
-    }
-
-    // Store new fingerprint
-    db()
-      .query(
-        "INSERT INTO kv_meta (key, value) VALUES (?, ?) ON CONFLICT(key) DO UPDATE SET value = ?",
-      )
-      .run(EMBEDDING_CONFIG_KEY, current, current);
-
-    return true;
-  };
-
-  return databaseInTransaction(db())
-    ? withSavepoint("embedding_config_reconcile", reconcile)
-    : withTransaction(reconcile);
-}
-
-/** Count blob-layout embeddings across all four tables (config-change logging). */
-function countBlobEmbeddings(): number {
-  const n = (sql: string) => (db().query(sql).get() as { n: number }).n;
-  return (
-    n(
-      "SELECT COUNT(*) as n FROM knowledge_current WHERE embedding IS NOT NULL",
-    ) +
-    n("SELECT COUNT(*) as n FROM distillations WHERE embedding IS NOT NULL") +
-    n(
-      "SELECT COUNT(*) as n FROM temporal_messages WHERE embedding IS NOT NULL",
-    ) +
-    n("SELECT COUNT(*) as n FROM entities WHERE embedding IS NOT NULL")
-  );
-}
-
-/** Count vec0-layout embeddings across all four `vec0` tables. Only reached on a
- *  capable runtime (degraded short-circuits earlier). */
-function countVec0Embeddings(): number {
-  const n = (sql: string) => (db().query(sql).get() as { n: number }).n;
-  return (
-    n("SELECT COUNT(*) as n FROM knowledge_vec") +
-    n("SELECT COUNT(*) as n FROM distillation_vec") +
-    n("SELECT COUNT(*) as n FROM temporal_vec") +
-    n("SELECT COUNT(*) as n FROM entity_vec")
-  );
-}
-
-/** The four embedding-bearing logical tables, in cutover order. */
-const EMBEDDING_TABLES: readonly EmbeddingTable[] = [
-  "knowledge",
-  "entities",
-  "distillations",
-  "temporal",
-];
-
-/**
- * One-time blob→vec0 cutover. NO-OP unless sqlite-vec is loadable on this
- * runtime AND the DB is still in blob layout (so an already-vec0 DB, or an
- * incapable runtime, skips). Idempotent / resumable: each table's
- * copy-then-drop is gated on the base `embedding` column still existing, so a
- * crash mid-cutover re-runs only the unfinished tables (never reads a dropped
- * column — the v55 boot-loop lesson). The storage mode flips LAST, only once
- * every table's blobs have been relocated and dropped.
- */
-export function maybeCutoverToVec0(): void {
-  if (!isVecAvailable()) return;
-
-  if (readStorageMode(db()) === "blob") {
-    const dim = config().search.embeddings.dimensions;
-    ensureVec0Store(db(), dim);
-    // Relocate every existing blob into vec0 BEFORE flipping the mode. The copy
-    // is idempotent (INSERT OR REPLACE) and does NOT drop anything, so a crash
-    // here leaves mode="blob" with the base columns still INTACT — the copy
-    // simply re-runs next startup. 🔴 INVARIANT: columns are dropped only AFTER
-    // the flip below, so mode==="blob" always implies the embedding columns
-    // still exist; no blob-mode query can ever read a half-dropped column (the
-    // v55 boot-loop hazard).
-    let staleSkipped = 0;
-    for (const table of EMBEDDING_TABLES) {
-      if (embeddingColumnExists(db(), table))
-        staleSkipped += copyBlobsToVec0(db(), table, dim);
-    }
-    // Flip once vec0 is fully populated and authoritative.
-    setStorageMode(db(), "vec0");
-    // Arm the temporal re-chunk walk so backfillTemporalEmbeddings definitely
-    // runs this startup and re-embeds every row skipped above (plus every legacy
-    // single-vector row) at the correct dimension. This is a no-op today — the
-    // done flag can only be latched from INSIDE a vec0-mode run of that walk, so
-    // a machine transitioning from blob mode never has it set — but calling it at
-    // the exact blob->vec0 transition hard-guards that invariant against future
-    // refactors and self-documents that the walk is (re)armed here.
-    resetTemporalRechunkProgress();
-    if (staleSkipped > 0) {
-      // Rare corpus corruption (blobs written under a different dimension). The
-      // rows were skipped from the copy and will be re-embedded at `dim` by the
-      // backfills below; surface it so an operator can see it happened.
-      log.notice(
-        `vec0 cutover skipped ${staleSkipped} stale-dimension embedding blob(s) (not ${dim}-dim / ${dim * 4} bytes); they will be re-embedded by the startup backfills`,
-      );
-    }
-    log.info(`vec0 storage cutover complete (dim=${dim})`);
-  }
-
-  // Reclaim: drop any leftover base embedding columns. Runs STRICTLY in vec0
-  // mode (mode never reverts to blob), so no blob-mode reader can observe a
-  // half-dropped column. Presence-aware + idempotent → resumable across a crash
-  // mid-drop (the next startup finishes the remaining columns).
-  if (readStorageMode(db()) === "vec0") {
-    let droppedAny = false;
-    for (const table of EMBEDDING_TABLES) {
-      if (embeddingColumnExists(db(), table)) {
-        dropEmbeddingColumn(db(), table);
-        droppedAny = true;
-      }
-    }
-    if (droppedAny) {
-      try {
-        // Best-effort: return the freed pages (notably ~320MB of temporal
-        // vectors) to the OS. No-op unless auto_vacuum is on.
-        db().query("PRAGMA incremental_vacuum").run();
-      } catch {
-        // ignore — space is already reclaimed within the DB file by DROP COLUMN.
-      }
-    }
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Startup backfill — single entry point for all hosts
-// ---------------------------------------------------------------------------
-
-/**
- * Delay before the startup backfill begins, so the host's HTTP server has
- * a clear window to answer the first wave of requests (web UI shell load,
- * terminal session-connect handshake) before the embedding worker starts
- * competing for CPU. With inference off the main thread the event loop
- * isn't blocked, but the worker still consumes a CPU core — a short delay
- * avoids contention during the first-connect burst.
- */
-const STARTUP_BACKFILL_DELAY_MS = 2_000;
-
-/**
- * Run all embedding backfills and log coverage stats.
- *
- * This is the canonical entry point that every host adapter (OpenCode, Pi,
- * future ACP) should call once during init. It:
- *   1. Waits a short grace period so first-connect HTTP requests can finish
- *   2. Detects config changes (provider swap) and clears stale embeddings
- *   3. Backfills knowledge entries missing embeddings
- *   4. Backfills non-archived distillations missing embeddings
- *   5. Logs a one-line coverage summary to stderr (always visible, not gated)
- *
- * Fire-and-forget: callers should `.catch()` — embedding failures must not
- * block plugin initialization.
- */
-/** Outcome of a startup backfill pass, returned for host-side instrumentation
- *  (the gateway wraps the call in a Sentry span). @loreai/core stays Sentry-free. */
-export interface BackfillStats {
-  pendingKnowledge: number;
-  pendingDistillations: number;
-  knowledgeEmbedded: number;
-  distillationEmbedded: number;
-  entityEmbedded: number;
-  knowledgeTotal: number;
-  knowledgeWithEmbedding: number;
-  distillationTotal: number;
-  distillationWithEmbedding: number;
-  temporalRechunked: number;
-}
-
-function emptyBackfillStats(): BackfillStats {
-  return {
-    pendingKnowledge: 0,
-    pendingDistillations: 0,
-    knowledgeEmbedded: 0,
-    distillationEmbedded: 0,
-    entityEmbedded: 0,
-    knowledgeTotal: 0,
-    knowledgeWithEmbedding: 0,
-    distillationTotal: 0,
-    distillationWithEmbedding: 0,
-    temporalRechunked: 0,
-  };
-}
-
-/** Host-supplied knobs for {@link runStartupBackfill}. */
-export interface BackfillOptions {
-  /**
-   * Idle-gate for the heavy, resumable temporal re-chunk walk. Consulted before
-   * each row; while it returns `true` the walk parks (re-polling) and resumes
-   * once it clears. `@loreai/core` has no view of host activity, so the host
-   * supplies the policy — the gateway pauses while background work is paused
-   * (circuit breaker) or a session was active within the idle window. Progress
-   * is durable per-row, so a long park just resumes later from the last
-   * checkpoint. Only the temporal walk is gated; the small knowledge/
-   * distillation/entity backfills run unthrottled.
-   */
-  shouldPause?: () => boolean;
-}
-
-export async function runStartupBackfill(
-  opts: BackfillOptions = {},
-): Promise<BackfillStats> {
-  if (!isAvailable()) {
-    // Make the degraded state visible in the startup path — this early return
-    // was previously silent, so a consumer who omitted the optional
-    // local-embedding stack (#1026), or is on a remote provider without a key,
-    // had no startup signal that backfill (and vector recall) is off. Gate on
-    // `enabled` so a deliberate `search.embeddings.enabled: false` stays quiet.
-    // `isAvailable()` already emits the local-broken FTS-only line once; this is
-    // the startup-scoped, backfill-specific companion.
-    if (config().search.embeddings.enabled !== false) {
-      log.info(
-        "startup embedding backfill skipped — embeddings unavailable " +
-          "(recall will use FTS-only search)",
-      );
-    }
-    const stats = emptyBackfillStats();
-    // Durable temporal admission is provider-independent. Preserve recovery
-    // progress even while vector recall is temporarily FTS-only.
-    stats.temporalRechunked = await backfillTemporalEmbeddings({
-      shouldPause: opts.shouldPause,
-    });
-    return stats;
-  }
-
-  // Handle an embedding-config change, then attempt the one-time blob→vec0
-  // cutover (both no-ops in the steady state). Order matters: a config change
-  // clears stale blobs BEFORE the cutover relocates the survivors, so the vec0
-  // tables are never seeded with vectors from a since-changed model/dimension.
-  checkConfigChange();
-  maybeCutoverToVec0();
-
-  const mode = readStorageMode(db());
-
-  // A vec0-store DB on a runtime that cannot load sqlite-vec: the blob columns
-  // are gone and the vec0 tables are unreadable. No backfill is possible — vector
-  // recall degrades to empty (FTS still answers) and re-converges when the DB is
-  // next opened on a capable runtime.
-  if (resolveReadMode(mode, isVecAvailable()) === "degraded") {
-    log.warn(
-      "vec0 storage but sqlite-vec unavailable — skipping embedding backfill " +
-        "(vector recall is FTS-only until reopened on a capable runtime)",
-    );
-    return emptyBackfillStats();
-  }
-
-  // Surface backlog up-front so a slow startup is self-explanatory in logs.
-  // Counts use the same mode-aware predicates the backfill loops use, so the
-  // two numbers always match what we're about to do. (In vec0 mode the blob
-  // column is gone — "pending" means a base row absent from the vec0 index.)
-  const pendingKnowledge = (
-    db()
-      .query(
-        `SELECT COUNT(*) as n FROM knowledge_current WHERE ${missingEmbeddingSql("knowledge", mode)} AND confidence > 0.2`,
-      )
-      .get() as { n: number }
-  ).n;
-  const pendingDistillations = (
-    db()
-      .query(
-        `SELECT COUNT(*) as n FROM distillations WHERE ${missingEmbeddingSql("distillations", mode)} AND archived = 0 AND observations != ''`,
-      )
-      .get() as { n: number }
-  ).n;
-
-  if (pendingKnowledge + pendingDistillations > 0) {
-    log.info(
-      `embedding backfill scheduled: ${pendingKnowledge} knowledge + ` +
-        `${pendingDistillations} distillations pending — starting in ` +
-        `${STARTUP_BACKFILL_DELAY_MS / 1000}s, batches yield between calls ` +
-        `(host stays responsive)`,
-    );
-    await new Promise<void>((r) => setTimeout(r, STARTUP_BACKFILL_DELAY_MS));
-  }
-
-  const knowledgeEmbedded = await backfillEmbeddings();
-  const distillationEmbedded = await backfillDistillationEmbeddings();
-  const entityEmbedded = await backfillEntityEmbeddings();
-  // Re-chunk pre-multi-vector temporal survivors into the vec0 layout. Resumable
-  // + done-flagged, so this is the heavy walk only on the first vec0 run (and
-  // again after a config change); a no-op in blob mode and once converged. Idle-
-  // gated (opts.shouldPause) so it yields the shared embed pool to live traffic.
-  const temporalConnection = db();
-  const temporalRechunked = await backfillTemporalEmbeddings({
-    shouldPause: opts.shouldPause,
-  });
-  // The walk may stop after shutdown. Do not reopen storage for GC or coverage
-  // stats, or use a successor connection that belongs to a different startup.
-  if (!isCurrentDatabase(temporalConnection)) {
-    return {
-      ...emptyBackfillStats(),
-      pendingKnowledge,
-      pendingDistillations,
-      knowledgeEmbedded,
-      distillationEmbedded,
-      entityEmbedded,
-      temporalRechunked,
-    };
-  }
-
-  // Orphan discovery belongs to the host's idle read-worker maintenance.
-  // Never scan the full vector corpus on the startup writer (#1681).
-
-  // Coverage stats — always log to stderr so the problem is visible.
-  const kTotal = (
-    db()
-      .query(
-        "SELECT COUNT(*) as n FROM knowledge_current WHERE confidence > 0.2",
-      )
-      .get() as { n: number }
-  ).n;
-  const kWithEmb = (
-    db()
-      .query(
-        `SELECT COUNT(*) as n FROM knowledge_current WHERE ${hasEmbeddingSql("knowledge", mode)} AND confidence > 0.2`,
-      )
-      .get() as { n: number }
-  ).n;
-  const dTotal = (
-    db()
-      .query(
-        "SELECT COUNT(*) as n FROM distillations WHERE archived = 0 AND observations != ''",
-      )
-      .get() as { n: number }
-  ).n;
-  const dWithEmb = (
-    db()
-      .query(
-        // Mirror dTotal's predicate (incl. observations != '') so the coverage
-        // numerator is always a subset of the denominator (never reads "11/10").
-        `SELECT COUNT(*) as n FROM distillations WHERE ${hasEmbeddingSql("distillations", mode)} AND archived = 0 AND observations != ''`,
-      )
-      .get() as { n: number }
-  ).n;
-
-  const parts: string[] = [];
-  // Lead with the storage mode + native availability so silent degradation is
-  // visible at a glance: `storage_mode=vec0 vec=off` means this DB cut over to
-  // vec0-only storage but sqlite-vec did not load here, so vector recall is
-  // FTS-only until reopened on a capable runtime.
-  parts.push(`storage_mode=${mode} vec=${isVecAvailable() ? "on" : "off"}`);
-  if (
-    knowledgeEmbedded > 0 ||
-    distillationEmbedded > 0 ||
-    entityEmbedded > 0 ||
-    temporalRechunked > 0
-  ) {
-    parts.push(
-      `backfilled ${knowledgeEmbedded} knowledge + ${distillationEmbedded} distillations + ${entityEmbedded} entities + ${temporalRechunked} temporal re-chunked`,
-    );
-  }
-  parts.push(
-    `coverage: knowledge ${kWithEmb}/${kTotal}, distillations ${dWithEmb}/${dTotal}`,
-  );
-  log.info(`embedding startup: ${parts.join("; ")}`);
-
-  return {
-    pendingKnowledge,
-    pendingDistillations,
-    knowledgeEmbedded,
-    distillationEmbedded,
-    entityEmbedded,
-    knowledgeTotal: kTotal,
-    knowledgeWithEmbedding: kWithEmb,
-    distillationTotal: dTotal,
-    distillationWithEmbedding: dWithEmb,
-    temporalRechunked,
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Backfill — knowledge
-// ---------------------------------------------------------------------------
-
-/**
- * Maximum chunk size for backfill embed requests. Each chunk becomes a
- * separate message to the embedding worker. Keeping chunks small gives
- * the worker's priority queue natural gaps to interleave high-priority
- * recall queries between backfill batches.
- */
-const MAX_BACKFILL_CHUNK = 8;
-
-/**
- * Maximum total "token area" (batch_size × max_sequence_length) per
- * backfill batch. ONNX runtime pads all texts to the longest sequence,
- * so the peak tensor size is proportional to this product. A budget of
- * 4096 tokens allows e.g. 8 × 512-token texts, or 2 × 2048-token texts.
- * Prevents OOM on batches with long distillation observations (~4000+
- * chars) that were blowing up at fixed batch sizes.
- */
-const MAX_BATCH_TOKEN_AREA = 4096;
-
-/**
- * Rough chars-per-token ratio for budget estimation. Nomic v1.5 uses a
- * WordPiece tokenizer; English text averages ~4 chars/token.
- */
-const CHARS_PER_TOKEN = 4;
-
-/**
- * Partition `rows` into batches that respect both MAX_BACKFILL_CHUNK and
- * MAX_BATCH_TOKEN_AREA. Each batch's estimated token area is
- * `batch.length × max_tokens_in_batch`. We greedily add rows until the
- * next row would push the area over budget.
- */
-function nextBatch<T extends { text: string }>(rows: T[], start: number): T[] {
-  const batch: T[] = [];
-  let maxTokens = 0;
-
-  for (
-    let i = start;
-    i < rows.length && batch.length < MAX_BACKFILL_CHUNK;
-    i++
-  ) {
-    const estTokens = Math.ceil(rows[i].text.length / CHARS_PER_TOKEN);
-    const newMax = Math.max(maxTokens, estTokens);
-    const newArea = (batch.length + 1) * newMax;
-
-    if (batch.length > 0 && newArea > MAX_BATCH_TOKEN_AREA) break;
-
-    batch.push(rows[i]);
-    maxTokens = newMax;
-  }
-
-  return batch;
-}
-
-/**
- * Embed all knowledge entries that are missing embeddings.
- * Called by `runStartupBackfill()`.
- * Also handles config changes: if provider/model/dimensions changed, clears
- * stale embeddings first, then re-embeds all entries.
- * When `options.signal` aborts or `options.deadlineMs` elapses, stops before
- * posting another batch and throws {@link EmbeddingAbortError}.
- * Returns the number of entries embedded.
- */
-export async function backfillEmbeddings(
-  options: EmbeddingOperationOptions = {},
-): Promise<number> {
-  const guard = createEmbeddingAbortGuard("knowledge-backfill", options);
-  throwIfEmbeddingAborted(guard);
-
-  // Detect config changes and clear stale embeddings
-  checkConfigChange();
-  throwIfEmbeddingAborted(guard);
-
-  const provider = getProvider();
-  if (!provider) {
-    throwIfEmbeddingAborted(guard);
-    return 0;
-  }
-
-  const mode = readStorageMode(db());
-  const rows = db()
-    .query(
-      `SELECT id, title, content FROM knowledge_current WHERE ${missingEmbeddingSql("knowledge", mode)} AND confidence > 0.2`,
-    )
-    .all() as Array<{ id: string; title: string; content: string }>;
-
-  throwIfEmbeddingAborted(guard);
-  if (!rows.length) return 0;
-
-  // Pre-compute text for token-budget batching
-  const items = rows.map((r) => ({ ...r, text: `${r.title}\n${r.content}` }));
-
-  let embedded = 0;
-  let i = 0;
-
-  while (i < items.length) {
-    // Check before selecting/posting each batch. If cancellation arrived while
-    // the previous batch settled, no new worker/API work is scheduled.
-    throwIfEmbeddingAborted(guard);
-    const batch = nextBatch(items, i);
-    i += batch.length;
-
-    try {
-      const vectors = await awaitEmbeddingOperation(
-        embed(
-          batch.map((b) => b.text),
-          "document",
-        ),
-        guard,
-      );
-      throwIfEmbeddingAborted(guard);
-
-      for (let j = 0; j < batch.length; j++) {
-        storeEmbedding(db(), "knowledge", batch[j].id, vectors[j]);
-        embedded++;
-      }
-    } catch (err) {
-      // Provider shutdown / unavailability is expected graceful degradation,
-      // not a bug — check before log.error so captureException doesn't fire
-      // and create Sentry noise (LOREAI-GATEWAY-Q).
-      if (err instanceof EmbeddingAbortError) throw err;
-      if (err instanceof EmbeddingQueueCapacityError) {
-        log.info("embedding backfill stopped: queue saturated");
-        break;
-      }
-      if (err instanceof LocalProviderUnavailableError) {
-        log.info("embedding backfill stopped: provider unavailable");
-        break;
-      }
-      log.error(
-        `embedding backfill batch failed (${batch.length} items):`,
-        err,
-      );
-    }
-    // No yieldToEventLoop() needed — embed() is truly async (worker thread).
-  }
-
-  if (embedded > 0) {
-    log.info(`embedded ${embedded} knowledge entries`);
-  }
-  return embedded;
-}
-
-// ---------------------------------------------------------------------------
-// Backfill — distillations
-// ---------------------------------------------------------------------------
-
-/**
- * Embed all non-archived distillations that are missing embeddings.
- * Called on startup alongside knowledge backfill.
- * Returns the number of distillations embedded.
- */
-export async function backfillDistillationEmbeddings(): Promise<number> {
-  const provider = getProvider();
-  if (!provider) return 0;
-
-  const mode = readStorageMode(db());
-  const rows = db()
-    .query(
-      `SELECT id, observations FROM distillations WHERE ${missingEmbeddingSql("distillations", mode)} AND archived = 0 AND observations != ''`,
-    )
-    .all() as Array<{ id: string; observations: string }>;
-
-  if (!rows.length) return 0;
-
-  let embedded = 0;
-
-  // Progress logging: heartbeat every PROGRESS_INTERVAL embedded so a long
-  // backfill (e.g. 1000+ pending after a model change) doesn't look
-  // like a silent hang. Without this, only the final tally was logged.
-  const PROGRESS_INTERVAL = 256;
-  let nextProgressAt = PROGRESS_INTERVAL;
-
-  // Pre-compute text for token-budget batching
-  const items = rows.map((r) => ({ ...r, text: r.observations }));
-  let i = 0;
-
-  while (i < items.length) {
-    const batch = nextBatch(items, i);
-    i += batch.length;
-
-    try {
-      const vectors = await embed(
-        batch.map((b) => b.text),
-        "document",
-      );
-
-      for (let j = 0; j < batch.length; j++) {
-        storeEmbedding(db(), "distillations", batch[j].id, vectors[j]);
-        embedded++;
-      }
-    } catch (err) {
-      // Provider shutdown / unavailability is expected graceful degradation,
-      // not a bug — check before log.error so captureException doesn't fire
-      // and create Sentry noise (LOREAI-GATEWAY-Q).
-      if (err instanceof EmbeddingQueueCapacityError) {
-        log.info("distillation embedding backfill stopped: queue saturated");
-        break;
-      }
-      if (err instanceof LocalProviderUnavailableError) {
-        log.info(
-          "distillation embedding backfill stopped: provider unavailable",
-        );
-        break;
-      }
-      log.error(
-        `distillation embedding backfill batch failed (${batch.length} items):`,
-        err,
-      );
-    }
-
-    if (embedded >= nextProgressAt) {
-      log.info(`embedding distillations: ${embedded}/${rows.length}…`);
-      nextProgressAt = embedded + PROGRESS_INTERVAL;
-    }
-    // No yieldToEventLoop() needed — embed() is truly async (worker thread).
-  }
-
-  if (embedded > 0) {
-    log.info(`embedded ${embedded} distillations`);
-  }
-  return embedded;
-}
-
-// ---------------------------------------------------------------------------
-// Backfill — entities
-// ---------------------------------------------------------------------------
-
-/**
- * Embed all entities that are missing embeddings. Composite text is the
- * canonical name plus all alias values. Called on startup alongside knowledge
- * and distillation backfill, and as a preflight before entity dedup so
- * similarity comparisons have vectors to work with. Returns the count embedded.
- */
-export async function backfillEntityEmbeddings(): Promise<number> {
-  const provider = getProvider();
-  if (!provider) return 0;
-
-  // Bun's bun:sqlite rejects GROUP_CONCAT(DISTINCT col, separator) with two
-  // arguments when DISTINCT is used ("DISTINCT aggregates must have exactly
-  // one argument"). Use a subquery to deduplicate aliases instead.
-  const mode = readStorageMode(db());
-  const rows = db()
-    .query(
-      `SELECT e.id AS id, e.canonical_name AS canonical_name,
-              (SELECT GROUP_CONCAT(da.alias_value, ' ')
-               FROM (SELECT DISTINCT alias_value FROM entity_aliases WHERE entity_id = e.id) da
-              ) AS aliases
-       FROM entities e
-       WHERE ${missingEmbeddingSql("entities", mode, "e")}`,
-    )
-    .all() as Array<{
-    id: string;
-    canonical_name: string;
-    aliases: string | null;
-  }>;
-
-  if (!rows.length) return 0;
-
-  // Pre-compute text for token-budget batching. Canonical name is also stored
-  // as a "name" alias, so it may already appear in `aliases` — harmless for the
-  // embedding (duplicate surface form), and entity names are short.
-  const items = rows.map((r) => ({
-    ...r,
-    text: `${r.canonical_name} ${r.aliases ?? ""}`.trim(),
-  }));
-
-  let embedded = 0;
-  let i = 0;
-
-  while (i < items.length) {
-    const batch = nextBatch(items, i);
-    i += batch.length;
-
-    try {
-      const vectors = await embed(
-        batch.map((b) => b.text),
-        "document",
-      );
-
-      for (let j = 0; j < batch.length; j++) {
-        storeEmbedding(db(), "entities", batch[j].id, vectors[j]);
-        embedded++;
-      }
-    } catch (err) {
-      // Provider shutdown / unavailability is expected graceful degradation,
-      // not a bug — check before log.error so captureException doesn't fire
-      // and create Sentry noise (LOREAI-GATEWAY-Q).
-      if (err instanceof EmbeddingQueueCapacityError) {
-        log.info("entity embedding backfill stopped: queue saturated");
-        break;
-      }
-      if (err instanceof LocalProviderUnavailableError) {
-        log.info("entity embedding backfill stopped: provider unavailable");
-        break;
-      }
-      log.error(
-        `entity embedding backfill batch failed (${batch.length} items):`,
-        err,
-      );
-    }
-    // No yieldToEventLoop() needed — embed() is truly async (worker thread).
-  }
-
-  if (embedded > 0) {
-    log.info(`embedded ${embedded} entities`);
-  }
-  return embedded;
-}
-
-// ---------------------------------------------------------------------------
-// Backfill — temporal messages (multi-vector re-chunk)
-// ---------------------------------------------------------------------------
-
-/** KV key: id of the last temporal message re-chunked by the backfill walk. */
-const TEMPORAL_RECHUNK_CURSOR_KEY = "lore:temporal_rechunk.cursor";
-/** KV key: "1" once the walk has reached the end of the corpus. */
-const TEMPORAL_RECHUNK_DONE_KEY = "lore:temporal_rechunk.done";
-/**
- * Rows fetched per page of the resumable temporal re-chunk walk. Bounds the
- * working-set memory (the corpus is 100k+ rows with large tool-output content)
- * and the redo cost of a crash mid-page.
- */
-const TEMPORAL_RECHUNK_PAGE = 256;
-/** Bound synchronous admission bursts, independently of provider duty settings.
- * The time budget is checked between rows; a single SQLite statement may exceed it. */
-const TEMPORAL_RECHUNK_YIELD_ROWS = 32;
-const TEMPORAL_RECHUNK_YIELD_MS = 8;
-const TEMPORAL_RECHUNK_ELIGIBLE_SQL = `length(CAST(content AS BLOB)) >= ${TEMPORAL_EMBEDDING_MIN_CONTENT_LENGTH}`;
-/** Legacy keys retained only so reset clears state written by pre-v85 builds. */
-const TEMPORAL_RECHUNK_ATTEMPTS_KEY = "lore:temporal_rechunk.attempts";
-const TEMPORAL_RECHUNK_INFLIGHT_KEY = "lore:temporal_rechunk.inflight";
-const TEMPORAL_RECHUNK_ROW_ATTEMPTS_KEY = "lore:temporal_rechunk.row_attempts";
-const TEMPORAL_RECHUNK_SKIP_KEY = "lore:temporal_rechunk.skip";
-const TEMPORAL_RECHUNK_MAX_ROWID_KEY = "lore:temporal_rechunk.max_rowid";
-/** Fence active walks across config resets, including resets by other processes. */
-const TEMPORAL_RECHUNK_EPOCH_KEY = "lore:temporal_rechunk.epoch";
-/**
- * Poll interval while the temporal walk is parked waiting for the host to go
- * idle. Short enough to resume promptly once traffic stops; a parked walk is
- * otherwise idle (one predicate call + one timer per tick).
- */
-const TEMPORAL_RECHUNK_PAUSE_POLL_MS = 250;
-
-/**
- * Park while the host-supplied idle-gate says to defer to live traffic. Returns
- * as soon as the gate clears (or immediately if no gate). A gate that throws is
- * treated as "not paused" — a buggy host predicate must never wedge the walk.
- */
-async function awaitBackfillIdle(
-  shouldPause: (() => boolean) | undefined,
-  isCurrent: () => boolean,
-): Promise<void> {
-  if (!shouldPause) return;
-  // Edge-triggered logging: because this call blocks until the host is idle,
-  // one busy stretch (however many rows it spans) yields exactly one park/resume
-  // pair — so a long park is visible in the logs instead of looking like a wedge.
-  let parked = false;
-  for (;;) {
-    if (!isCurrent()) return;
-    let paused = false;
-    try {
-      paused = shouldPause();
-    } catch {
-      break; // never let a throwing gate brick the walk
-    }
-    if (!paused) break;
-    if (!parked) {
-      parked = true;
-      log.info("temporal re-chunk parked — deferring to live traffic");
-    }
-    await new Promise<void>((resolve) => {
-      const t = setTimeout(resolve, TEMPORAL_RECHUNK_PAUSE_POLL_MS);
-      // Don't let a parked walk hold the process open on its own; the host's
-      // server keeps it alive, and if it doesn't, resuming next start is fine.
-      (t as { unref?: () => void }).unref?.();
-    });
-  }
-  if (parked) log.info("temporal re-chunk resumed");
-}
-
-/**
- * Reset durable scheduling progress so the next startup walks the corpus again.
- * Called when an embedding config change clears all vectors; the scheduler then
- * repopulates them from the queue under the new fingerprint.
- */
-export function resetTemporalRechunkProgress(): void {
-  withSavepoint("reset_temporal_rechunk", () => {
-    setKV(TEMPORAL_RECHUNK_EPOCH_KEY, randomUUID());
-    setKV(TEMPORAL_RECHUNK_DONE_KEY, "0");
-    setKV(TEMPORAL_RECHUNK_CURSOR_KEY, "");
-    setKV(TEMPORAL_RECHUNK_ATTEMPTS_KEY, "0");
-    setKV(TEMPORAL_RECHUNK_INFLIGHT_KEY, "");
-    setKV(TEMPORAL_RECHUNK_ROW_ATTEMPTS_KEY, "0");
-    setKV(TEMPORAL_RECHUNK_SKIP_KEY, "");
-    setKV(TEMPORAL_RECHUNK_MAX_ROWID_KEY, "0");
-  });
-}
-
-/**
- * Cumulative-progress line for the temporal re-chunk walk. `done`/`total` count
- * embeddable (>=50 char) messages across ALL runs, so the percentage keeps
- * climbing over restarts — unlike the per-process counters, which reset to zero
- * on every boot. `thisRun` is what the current invocation scheduled. A zero
- * `total` reads as 100% (nothing to do). The percentage is clamped to 100 and
- * rendered to one decimal. Exported for testing. See {@link
- * backfillTemporalEmbeddings} for how `done = baseDone + scanned` is derived.
- */
-export function formatTemporalRechunkProgress(
-  done: number,
-  total: number,
-  thisRun: number,
-): string {
-  const pct =
-    total > 0 ? Math.min(100, Math.round((done / total) * 1000) / 10) : 100;
-  return `temporal re-chunk: ${pct}% complete (${done}/${total} messages) · +${thisRun} scheduled this run`;
-}
-
-/**
- * Durably schedule existing temporal messages for the multi-vector vec0 layout.
- *
- * New messages are embedded part-aware at write time, but messages written
- * before the multi-vector policy — and every survivor of the blob→vec0 cutover,
- * which copies a single legacy vector as one `#0` chunk — still carry a single
- * vector. Distilled messages stay in the vector search path, so the legacy set
- * is the whole corpus. This walk only admits work; the bounded durable scheduler
- * is the sole temporal inference and vector-write owner.
- *
- * Properties:
- *  - vec0-only: in blob mode there is one vector per row regardless, so this is
- *    a no-op and does NOT latch the done flag — it runs after a future cutover.
- *  - resumable: each queue upsert, stale-vector invalidation, and cursor advance
- *    share one savepoint. An interrupted run resumes from the last admitted id.
- *  - finite snapshot: the first run persists the maximum eligible base rowid;
- *    live rows inserted later are owned only by the scheduler and never extend
- *    or get invalidated by this migration.
- *  - stable order, not chronological: temporal ids come from the upstream agent
- *    (not lore), so `id` may not be time-ordered. The walk only needs a stable,
- *    unique key — `id TEXT PRIMARY KEY`, `ORDER BY id ASC` (BINARY) — to visit
- *    every pre-existing row exactly once. New rows inserted below the cursor are
- *    already correct from write time, so skipping them is harmless.
- *  - provider-independent: admission never calls an embedding provider, so a
- *    worker crash or remote failure cannot poison or pin migration progress.
- *  - bounded memory: pages the scan instead of loading the whole corpus.
- *  - converges once: a done flag latches at the end of a clean walk; subsequent
- *    startups skip it. {@link resetTemporalRechunkProgress} re-arms it after a
- *    model/dimension change.
- *
- * Returns the number of messages durably scheduled in this invocation.
- */
-export async function backfillTemporalEmbeddings(
-  opts: { shouldPause?: () => boolean } = {},
-): Promise<number> {
-  // Multi-vector chunking only exists in vec0 mode. Skip WITHOUT latching done
-  // so a later cutover to vec0 still triggers the walk.
-  //
-  // 🔴 Ordering invariant: this vec0-mode guard MUST come BEFORE the done-flag
-  // check below. Because the flag can therefore only ever be latched from inside
-  // a vec0-mode run, a blob-mode run can never set it — which is what guarantees
-  // the first walk after a blob->vec0 cutover is always armed and re-embeds the
-  // rows that cutover skipped. (maybeCutoverToVec0 also explicitly re-arms the
-  // walk on cutover as belt-and-suspenders.) Do not reorder these two lines.
-  const connection = db();
-  const snapshot = withSavepoint("start_temporal_rechunk", () => {
-    if (readStorageMode(connection) !== "vec0") return null;
-    if (getKV(TEMPORAL_RECHUNK_DONE_KEY) === "1") return null;
-    const cursor = getKV(TEMPORAL_RECHUNK_CURSOR_KEY) ?? "";
-    const epoch = getKV(TEMPORAL_RECHUNK_EPOCH_KEY);
-    let maxRowid = Number(getKV(TEMPORAL_RECHUNK_MAX_ROWID_KEY) ?? "0");
-    if (!Number.isSafeInteger(maxRowid) || maxRowid <= 0) {
-      maxRowid = (
-        connection
-          .query(
-            `SELECT COALESCE(MAX(rowid), 0) AS n FROM temporal_messages WHERE ${TEMPORAL_RECHUNK_ELIGIBLE_SQL}`,
-          )
-          .get() as { n: number }
-      ).n;
-      setKV(TEMPORAL_RECHUNK_MAX_ROWID_KEY, String(maxRowid));
-    }
-    return { cursor, epoch, maxRowid };
-  });
-  if (!snapshot) return 0;
-  let { cursor } = snapshot;
-  const { maxRowid, epoch } = snapshot;
-  // Identity must be checked first: getKV/db() would reopen storage after close.
-  const isCurrent = () =>
-    isCurrentDatabase(connection) &&
-    getKV(TEMPORAL_RECHUNK_EPOCH_KEY) === epoch;
-  let scheduled = 0;
-  let scanned = 0;
-
-  // Up-front backlog so a long walk is explainable from the logs. The corpus is
-  // 100k+ rows and, under embed-pool contention, converges at single-digit
-  // rows/min — without this line a multi-hour (or, across restarts, multi-day)
-  // walk is completely invisible. Counts rows still to scan from the resume
-  // cursor; runs once per process (the walk is one-shot per config).
-  const backlog = (
-    db()
-      .query(
-        `SELECT COUNT(*) AS n FROM temporal_messages
-         WHERE id > ? AND rowid <= ? AND ${TEMPORAL_RECHUNK_ELIGIBLE_SQL}`,
-      )
-      .get(cursor, maxRowid) as { n: number }
-  ).n;
-  // Denominator for cumulative progress: ALL embeddable messages, not just the
-  // remaining backlog, so the heartbeat shows a percentage that keeps climbing
-  // across restarts instead of the per-process tally that resets to zero on every
-  // boot. `baseDone` is what prior runs already covered (rows at/below the resume
-  // cursor); the walk's SELECT uses the same byte-length eligibility predicate,
-  // so `scanned`
-  // counts embeddable rows and `baseDone + scanned` reaches exactly `total` on a
-  // clean pass (→ 100%). Gated on `backlog > 0`: when nothing remains the loop
-  // latches done without iterating, so `total`/`baseDone` are never read — no
-  // point paying for a second full-table COUNT on that path.
-  let total = 0;
-  let baseDone = 0;
-  if (backlog > 0) {
-    total = (
-      db()
-        .query(
-          `SELECT COUNT(*) AS n FROM temporal_messages WHERE rowid <= ? AND ${TEMPORAL_RECHUNK_ELIGIBLE_SQL}`,
-        )
-        .get(maxRowid) as { n: number }
-    ).n;
-    baseDone = Math.max(0, total - backlog);
-    const basePct = total > 0 ? Math.round((baseDone / total) * 1000) / 10 : 0;
-    log.info(
-      `temporal re-chunk: ${backlog} messages to scan (${baseDone}/${total} already done, ${basePct}%)${cursor ? ", resuming" : ""}`,
-    );
-  }
-
-  // Wall-clock heartbeat rather than a per-N-rows milestone: at the observed
-  // throughput a 1000-row milestone can be hours away, so a time cadence keeps
-  // the walk visible regardless of speed.
-  const PROGRESS_INTERVAL_MS = 30_000;
-  let lastProgressAt = Date.now();
-  let sliceStarted = performance.now();
-
-  for (;;) {
-    if (!isCurrent()) return scheduled;
-    // Keep the legacy walk aligned with scheduler eligibility.
-    const rows = db()
-      .query(
-        `SELECT id FROM temporal_messages
-         WHERE id > ? AND rowid <= ? AND ${TEMPORAL_RECHUNK_ELIGIBLE_SQL}
-         ORDER BY id ASC LIMIT ?`,
-      )
-      .all(cursor, maxRowid, TEMPORAL_RECHUNK_PAGE) as Array<{ id: string }>;
-
-    if (!rows.length) {
-      withSavepoint("finish_temporal_rechunk", () => {
-        if (!isCurrent()) return;
-        setKV(TEMPORAL_RECHUNK_DONE_KEY, "1");
-        setKV(TEMPORAL_RECHUNK_ATTEMPTS_KEY, "0");
-        setKV(TEMPORAL_RECHUNK_INFLIGHT_KEY, "");
-        setKV(TEMPORAL_RECHUNK_ROW_ATTEMPTS_KEY, "0");
-        setKV(TEMPORAL_RECHUNK_SKIP_KEY, "");
-      });
-      break;
-    }
-
-    for (const row of rows) {
-      // Admission performs synchronous SQLite work only. Keep the host gate so a
-      // one-time full-corpus walk still yields between rows under request load.
-      await awaitBackfillIdle(opts.shouldPause, isCurrent);
-      if (!isCurrentDatabase(connection)) return scheduled;
-      const admitted = withSavepoint("schedule_temporal_rechunk", () => {
-        if (!isCurrent()) return false;
-        const current = db()
-          .query(
-            "SELECT content FROM temporal_messages WHERE id = ? AND rowid <= ?",
-          )
-          .get(row.id, maxRowid) as { content: string } | null;
-        if (current) {
-          const enqueued = enqueueTemporalEmbedding(row.id, current.content);
-          if (enqueued) {
-            invalidateTemporalEmbedding(row.id);
-            scheduled++;
-          }
-        }
-        cursor = row.id;
-        scanned++;
-        setKV(TEMPORAL_RECHUNK_CURSOR_KEY, cursor);
-        return true;
-      });
-      if (!admitted) return scheduled;
-
-      if (Date.now() - lastProgressAt >= PROGRESS_INTERVAL_MS) {
-        log.info(
-          formatTemporalRechunkProgress(baseDone + scanned, total, scheduled),
-        );
-        lastProgressAt = Date.now();
-      }
-
-      // awaitBackfillIdle resolves immediately when the gate is clear. Awaiting
-      // it only yields to microtasks, starving HTTP/timers for the entire walk.
-      // Yield after SCANNED rows (including already-queued/deleted rows), with
-      // every queue/vector/cursor savepoint committed before other work runs.
-      if (
-        scanned % TEMPORAL_RECHUNK_YIELD_ROWS === 0 ||
-        performance.now() - sliceStarted >= TEMPORAL_RECHUNK_YIELD_MS
-      ) {
-        await new Promise<void>((resolve) => setImmediate(resolve));
-        if (!isCurrent()) return scheduled;
-        sliceStarted = performance.now();
-      }
-    }
-  }
-
-  if (scheduled > 0) {
-    log.info(
-      formatTemporalRechunkProgress(baseDone + scanned, total, scheduled),
-    );
-  }
-  return scheduled;
-}
