+++ packages/website/src/content/docs/docs/environment.md
@@ -96,7 +96,7 @@ Env vars override `.lore.json` for the same setting. To override a `.lore.json` 
 | `LORE_BACKFILL_CPU_DUTY` | Legacy temporal backfill duty setting (`LORE_BACKFILL_CPU_DUTY`) retained for configuration compatibility. Durable temporal admission no longer performs inference; the bounded scheduler controls embedding concurrency. |
 | `LORE_DISABLE_VEC` | LORE_DISABLE_VEC=1 forces the JS brute-force vector-search path. Useful as a production kill-switch if the native extension causes issues, and as a test seam for the JS fallback. Set before the first `db()` call — once attempted=true is sticky for the connection lifetime, the env var won't be re-read until resetVecState() runs (in close()). |
 | `LORE_DISABLE_VEC_WORKER` | Kill switch: force the in-process vector-search path, disabling the off-thread read-worker pool. Default-on escape hatch, not opt-in. |
-| `LORE_EMBED_POOL_SIZE` | Resolve the configured embedding-pool ceiling: `LORE_EMBED_POOL_SIZE` wins (the escape hatch — `=1` forces today's single worker), then `search.embeddings.embedPoolSize`, else `undefined` (memory-gated default). Read per-construction (not cached) so config reloads / env changes take effect on the next provider. Invalid values are ignored (fall through). |
+| `LORE_EMBED_POOL_SIZE` | Number of local ONNX embedding worker threads (each loads its own ~137MB model). Distinct from workerPoolSize (the cheap DB-only vector pool) because an embedding worker's memory footprint is far larger. Omit to let lore size the pool from available memory (1 on constrained hosts, up to 2 when there's headroom). Override with LORE_EMBED_POOL_SIZE. |
 | `LORE_NO_DB_TRACING` | LORE_NO_DB_TRACING=1 returns the raw connection instead of the query-tracing Proxy (disables automatic per-query DB spans). |
 | `LORE_VEC_SEARCH_TIMEOUT_MS` | LORE_VEC_SEARCH_TIMEOUT_MS overrides the per-request vector-search timeout (a positive integer in milliseconds; invalid or non-positive values are ignored). Defaults to 10000 (10s). On timeout, recall degrades to an empty result instead of re-running the O(n) scan on the main thread. |
